% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_activate_event_source}
\alias{eventbridge_activate_event_source}
\title{Activates a partner event source that has been deactivated}
\usage{
eventbridge_activate_event_source(Name)
}
\arguments{
\item{Name}{[required] The name of the partner event source to activate.}
}
\description{
Activates a partner event source that has been deactivated. Once
activated, your matching event bus will start receiving events from the
event source.
}
\details{
This operation is performed by AWS customers, not by SaaS partners.
}
\section{Request syntax}{
\preformatted{svc$activate_event_source(
  Name = "string"
)
}
}

\keyword{internal}
