% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_create_api_destination}
\alias{eventbridge_create_api_destination}
\title{Creates an API destination, which is an HTTP invocation endpoint
configured as a target for events}
\usage{
eventbridge_create_api_destination(
  Name,
  Description = NULL,
  ConnectionArn,
  InvocationEndpoint,
  HttpMethod,
  InvocationRateLimitPerSecond = NULL
)
}
\arguments{
\item{Name}{[required] The name for the API destination to create.}

\item{Description}{A description for the API destination to create.}

\item{ConnectionArn}{[required] The ARN of the connection to use for the API destination. The
destination endpoint must support the authorization type specified for
the connection.}

\item{InvocationEndpoint}{[required] The URL to the HTTP invocation endpoint for the API destination.}

\item{HttpMethod}{[required] The method to use for the request to the HTTP invocation endpoint.}

\item{InvocationRateLimitPerSecond}{The maximum number of requests per second to send to the HTTP invocation
endpoint.}
}
\description{
Creates an API destination, which is an HTTP invocation endpoint configured as a target for events.

See \url{https://paws-r.github.io/docs/eventbridge/create_api_destination.html} for full documentation.
}
\keyword{internal}
