% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_deactivate_event_source}
\alias{eventbridge_deactivate_event_source}
\title{You can use this operation to temporarily stop receiving events from the
specified partner event source}
\usage{
eventbridge_deactivate_event_source(Name)
}
\arguments{
\item{Name}{[required] The name of the partner event source to deactivate.}
}
\description{
You can use this operation to temporarily stop receiving events from the specified partner event source. The matching event bus is not deleted.

See \url{https://paws-r.github.io/docs/eventbridge/deactivate_event_source.html} for full documentation.
}
\keyword{internal}
