% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_delete_api_destination}
\alias{eventbridge_delete_api_destination}
\title{Deletes the specified API destination}
\usage{
eventbridge_delete_api_destination(Name)
}
\arguments{
\item{Name}{[required] The name of the destination to delete.}
}
\description{
Deletes the specified API destination.

See \url{https://paws-r.github.io/docs/eventbridge/delete_api_destination.html} for full documentation.
}
\keyword{internal}
