% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_delete_archive}
\alias{eventbridge_delete_archive}
\title{Deletes the specified archive}
\usage{
eventbridge_delete_archive(ArchiveName)
}
\arguments{
\item{ArchiveName}{[required] The name of the archive to delete.}
}
\value{
An empty list.
}
\description{
Deletes the specified archive.
}
\section{Request syntax}{
\preformatted{svc$delete_archive(
  ArchiveName = "string"
)
}
}

\keyword{internal}
