% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_delete_endpoint}
\alias{eventbridge_delete_endpoint}
\title{Delete an existing global endpoint}
\usage{
eventbridge_delete_endpoint(Name)
}
\arguments{
\item{Name}{[required] The name of the endpoint you want to delete. For example,
\code{"Name":"us-east-2-custom_bus_A-endpoint"}..}
}
\description{
Delete an existing global endpoint. For more information about global endpoints, see \href{https://docs.aws.amazon.com/eventbridge/latest/userguide/eb-global-endpoints.html}{Making applications Regional-fault tolerant with global endpoints and event replication} in the \emph{Amazon EventBridge User Guide} .

See \url{https://www.paws-r-sdk.com/docs/eventbridge_delete_endpoint/} for full documentation.
}
\keyword{internal}
