% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_describe_api_destination}
\alias{eventbridge_describe_api_destination}
\title{Retrieves details about an API destination}
\usage{
eventbridge_describe_api_destination(Name)
}
\arguments{
\item{Name}{[required] The name of the API destination to retrieve.}
}
\value{
A list with the following syntax:\preformatted{list(
  ApiDestinationArn = "string",
  Name = "string",
  Description = "string",
  ApiDestinationState = "ACTIVE"|"INACTIVE",
  ConnectionArn = "string",
  InvocationEndpoint = "string",
  HttpMethod = "POST"|"GET"|"HEAD"|"OPTIONS"|"PUT"|"PATCH"|"DELETE",
  InvocationRateLimitPerSecond = 123,
  CreationTime = as.POSIXct(
    "2015-01-01"
  ),
  LastModifiedTime = as.POSIXct(
    "2015-01-01"
  )
)
}
}
\description{
Retrieves details about an API destination.
}
\section{Request syntax}{
\preformatted{svc$describe_api_destination(
  Name = "string"
)
}
}

\keyword{internal}
