% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_describe_connection}
\alias{eventbridge_describe_connection}
\title{Retrieves details about a connection}
\usage{
eventbridge_describe_connection(Name)
}
\arguments{
\item{Name}{[required] The name of the connection to retrieve.}
}
\value{
A list with the following syntax:\preformatted{list(
  ConnectionArn = "string",
  Name = "string",
  Description = "string",
  ConnectionState = "CREATING"|"UPDATING"|"DELETING"|"AUTHORIZED"|"DEAUTHORIZED"|"AUTHORIZING"|"DEAUTHORIZING",
  StateReason = "string",
  AuthorizationType = "BASIC"|"OAUTH_CLIENT_CREDENTIALS"|"API_KEY",
  SecretArn = "string",
  AuthParameters = list(
    BasicAuthParameters = list(
      Username = "string"
    ),
    OAuthParameters = list(
      ClientParameters = list(
        ClientID = "string"
      ),
      AuthorizationEndpoint = "string",
      HttpMethod = "GET"|"POST"|"PUT",
      OAuthHttpParameters = list(
        HeaderParameters = list(
          list(
            Key = "string",
            Value = "string",
            IsValueSecret = TRUE|FALSE
          )
        ),
        QueryStringParameters = list(
          list(
            Key = "string",
            Value = "string",
            IsValueSecret = TRUE|FALSE
          )
        ),
        BodyParameters = list(
          list(
            Key = "string",
            Value = "string",
            IsValueSecret = TRUE|FALSE
          )
        )
      )
    ),
    ApiKeyAuthParameters = list(
      ApiKeyName = "string"
    ),
    InvocationHttpParameters = list(
      HeaderParameters = list(
        list(
          Key = "string",
          Value = "string",
          IsValueSecret = TRUE|FALSE
        )
      ),
      QueryStringParameters = list(
        list(
          Key = "string",
          Value = "string",
          IsValueSecret = TRUE|FALSE
        )
      ),
      BodyParameters = list(
        list(
          Key = "string",
          Value = "string",
          IsValueSecret = TRUE|FALSE
        )
      )
    )
  ),
  CreationTime = as.POSIXct(
    "2015-01-01"
  ),
  LastModifiedTime = as.POSIXct(
    "2015-01-01"
  ),
  LastAuthorizedTime = as.POSIXct(
    "2015-01-01"
  )
)
}
}
\description{
Retrieves details about a connection.
}
\section{Request syntax}{
\preformatted{svc$describe_connection(
  Name = "string"
)
}
}

\keyword{internal}
