% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_describe_event_source}
\alias{eventbridge_describe_event_source}
\title{This operation lists details about a partner event source that is shared
with your account}
\usage{
eventbridge_describe_event_source(Name)
}
\arguments{
\item{Name}{[required] The name of the partner event source to display the details of.}
}
\description{
This operation lists details about a partner event source that is shared with your account.

See \url{https://www.paws-r-sdk.com/docs/eventbridge_describe_event_source/} for full documentation.
}
\keyword{internal}
