% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_describe_replay}
\alias{eventbridge_describe_replay}
\title{Retrieves details about a replay}
\usage{
eventbridge_describe_replay(ReplayName)
}
\arguments{
\item{ReplayName}{[required] The name of the replay to retrieve.}
}
\description{
Retrieves details about a replay. Use \code{DescribeReplay} to determine the
progress of a running replay. A replay processes events to replay based
on the time in the event, and replays them using 1 minute intervals. If
you use \code{StartReplay} and specify an \code{EventStartTime} and an
\code{EventEndTime} that covers a 20 minute time range, the events are
replayed from the first minute of that 20 minute range first. Then the
events from the second minute are replayed. You can use \code{DescribeReplay}
to determine the progress of a replay. The value returned for
\code{EventLastReplayedTime} indicates the time within the specified time
range associated with the last event replayed.
}
\section{Request syntax}{
\preformatted{svc$describe_replay(
  ReplayName = "string"
)
}
}

\keyword{internal}
