% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_list_archives}
\alias{eventbridge_list_archives}
\title{Lists your archives}
\usage{
eventbridge_list_archives(
  NamePrefix = NULL,
  EventSourceArn = NULL,
  State = NULL,
  NextToken = NULL,
  Limit = NULL
)
}
\arguments{
\item{NamePrefix}{A name prefix to filter the archives returned. Only archives with name
that match the prefix are returned.}

\item{EventSourceArn}{The ARN of the event source associated with the archive.}

\item{State}{The state of the archive.}

\item{NextToken}{The token returned by a previous call, which you can use to retrieve the
next set of results.

The value of \code{nextToken} is a unique pagination token for each page. To
retrieve the next page of results, make the call again using the
returned token. Keep all other arguments unchanged.

Using an expired pagination token results in an \verb{HTTP 400 InvalidToken}
error.}

\item{Limit}{The maximum number of results to return.}
}
\description{
Lists your archives. You can either list all the archives or you can provide a prefix to match to the archive names. Filter parameters are exclusive.

See \url{https://www.paws-r-sdk.com/docs/eventbridge_list_archives/} for full documentation.
}
\keyword{internal}
