% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_list_connections}
\alias{eventbridge_list_connections}
\title{Retrieves a list of connections from the account}
\usage{
eventbridge_list_connections(
  NamePrefix = NULL,
  ConnectionState = NULL,
  NextToken = NULL,
  Limit = NULL
)
}
\arguments{
\item{NamePrefix}{A name prefix to filter results returned. Only connections with a name
that starts with the prefix are returned.}

\item{ConnectionState}{The state of the connection.}

\item{NextToken}{The token returned by a previous call, which you can use to retrieve the
next set of results.

The value of \code{nextToken} is a unique pagination token for each page. To
retrieve the next page of results, make the call again using the
returned token. Keep all other arguments unchanged.

Using an expired pagination token results in an \verb{HTTP 400 InvalidToken}
error.}

\item{Limit}{The maximum number of connections to return.}
}
\description{
Retrieves a list of connections from the account.

See \url{https://www.paws-r-sdk.com/docs/eventbridge_list_connections/} for full documentation.
}
\keyword{internal}
