% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_list_partner_event_source_accounts}
\alias{eventbridge_list_partner_event_source_accounts}
\title{An SaaS partner can use this operation to display the Amazon Web
Services account ID that a particular partner event source name is
associated with}
\usage{
eventbridge_list_partner_event_source_accounts(
  EventSourceName,
  NextToken = NULL,
  Limit = NULL
)
}
\arguments{
\item{EventSourceName}{[required] The name of the partner event source to display account information
about.}

\item{NextToken}{The token returned by a previous call, which you can use to retrieve the
next set of results.

The value of \code{nextToken} is a unique pagination token for each page. To
retrieve the next page of results, make the call again using the
returned token. Keep all other arguments unchanged.

Using an expired pagination token results in an \verb{HTTP 400 InvalidToken}
error.}

\item{Limit}{Specifying this limits the number of results returned by this operation.
The operation also returns a NextToken which you can use in a subsequent
operation to retrieve the next set of results.}
}
\description{
An SaaS partner can use this operation to display the Amazon Web Services account ID that a particular partner event source name is associated with. This operation is not used by Amazon Web Services customers.

See \url{https://www.paws-r-sdk.com/docs/eventbridge_list_partner_event_source_accounts/} for full documentation.
}
\keyword{internal}
