% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_put_partner_events}
\alias{eventbridge_put_partner_events}
\title{This is used by SaaS partners to write events to a customer's partner
event bus}
\usage{
eventbridge_put_partner_events(Entries)
}
\arguments{
\item{Entries}{[required] The list of events to write to the event bus.}
}
\description{
This is used by SaaS partners to write events to a customer's partner event bus. Amazon Web Services customers do not use this operation.

See \url{https://paws-r.github.io/docs/eventbridge/put_partner_events.html} for full documentation.
}
\keyword{internal}
