% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_put_targets}
\alias{eventbridge_put_targets}
\title{Adds the specified targets to the specified rule, or updates the targets
if they are already associated with the rule}
\usage{
eventbridge_put_targets(Rule, EventBusName = NULL, Targets)
}
\arguments{
\item{Rule}{[required] The name of the rule.}

\item{EventBusName}{The name or ARN of the event bus associated with the rule. If you omit
this, the default event bus is used.}

\item{Targets}{[required] The targets to update or add to the rule.}
}
\description{
Adds the specified targets to the specified rule, or updates the targets if they are already associated with the rule.

See \url{https://paws-r.github.io/docs/eventbridge/put_targets.html} for full documentation.
}
\keyword{internal}
