% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_tag_resource}
\alias{eventbridge_tag_resource}
\title{Assigns one or more tags (key-value pairs) to the specified EventBridge
resource}
\usage{
eventbridge_tag_resource(ResourceARN, Tags)
}
\arguments{
\item{ResourceARN}{[required] The ARN of the EventBridge resource that you're adding tags to.}

\item{Tags}{[required] The list of key-value pairs to associate with the resource.}
}
\description{
Assigns one or more tags (key-value pairs) to the specified EventBridge resource. Tags can help you organize and categorize your resources. You can also use them to scope user permissions by granting a user permission to access or change only resources with certain tag values. In EventBridge, rules and event buses can be tagged.

See \url{https://paws-r.github.io/docs/eventbridge/tag_resource.html} for full documentation.
}
\keyword{internal}
