% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_test_event_pattern}
\alias{eventbridge_test_event_pattern}
\title{Tests whether the specified event pattern matches the provided event}
\usage{
eventbridge_test_event_pattern(EventPattern, Event)
}
\arguments{
\item{EventPattern}{[required] The event pattern. For more information, see \href{https://docs.aws.amazon.com/eventbridge/latest/userguide/eb-events.html}{Events and Event Patterns}
in the \emph{Amazon EventBridge User Guide}.}

\item{Event}{[required] The event, in JSON format, to test against the event pattern. The JSON
must follow the format specified in \href{https://docs.aws.amazon.com/eventbridge/latest/userguide/eb-events.html}{AWS Events},
and the following fields are mandatory:
\itemize{
\item \code{id}
\item \code{account}
\item \code{source}
\item \code{time}
\item \code{region}
\item \code{resources}
\item \code{detail-type}
}}
}
\value{
A list with the following syntax:\preformatted{list(
  Result = TRUE|FALSE
)
}
}
\description{
Tests whether the specified event pattern matches the provided event.

Most services in AWS treat : or / as the same character in Amazon
Resource Names (ARNs). However, EventBridge uses an exact match in event
patterns and rules. Be sure to use the correct ARN characters when
creating event patterns so that they match the ARN syntax in the event
you want to match.
}
\section{Request syntax}{
\preformatted{svc$test_event_pattern(
  EventPattern = "string",
  Event = "string"
)
}
}

\keyword{internal}
