% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_untag_resource}
\alias{eventbridge_untag_resource}
\title{Removes one or more tags from the specified EventBridge resource}
\usage{
eventbridge_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] The ARN of the EventBridge resource from which you are removing tags.}

\item{TagKeys}{[required] The list of tag keys to remove from the resource.}
}
\value{
An empty list.
}
\description{
Removes one or more tags from the specified EventBridge resource. In
Amazon EventBridge (CloudWatch Events, rules and event buses can be
tagged.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceARN = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
