% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_update_archive}
\alias{eventbridge_update_archive}
\title{Updates the specified archive}
\usage{
eventbridge_update_archive(
  ArchiveName,
  Description = NULL,
  EventPattern = NULL,
  RetentionDays = NULL,
  KmsKeyIdentifier = NULL
)
}
\arguments{
\item{ArchiveName}{[required] The name of the archive to update.}

\item{Description}{The description for the archive.}

\item{EventPattern}{The event pattern to use to filter events sent to the archive.}

\item{RetentionDays}{The number of days to retain events in the archive.}

\item{KmsKeyIdentifier}{The identifier of the KMS customer managed key for EventBridge to use,
if you choose to use a customer managed key to encrypt this archive. The
identifier can be the key Amazon Resource Name (ARN), KeyId, key alias,
or key alias ARN.

If you do not specify a customer managed key identifier, EventBridge
uses an Amazon Web Services owned key to encrypt the archive.

For more information, see \href{https://docs.aws.amazon.com/kms/latest/developerguide/viewing-keys.html}{Identify and view keys}
in the \emph{Key Management Service Developer Guide}.

If you have specified that EventBridge use a customer managed key for
encrypting the source event bus, we strongly recommend you also specify
a customer managed key for any archives for the event bus as well.

For more information, see \href{https://docs.aws.amazon.com/eventbridge/latest/userguide/encryption-archives.html}{Encrypting archives}
in the \emph{Amazon EventBridge User Guide}.}
}
\description{
Updates the specified archive.

See \url{https://www.paws-r-sdk.com/docs/eventbridge_update_archive/} for full documentation.
}
\keyword{internal}
