% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_update_connection}
\alias{eventbridge_update_connection}
\title{Updates settings for a connection}
\usage{
eventbridge_update_connection(
  Name,
  Description = NULL,
  AuthorizationType = NULL,
  AuthParameters = NULL,
  InvocationConnectivityParameters = NULL,
  KmsKeyIdentifier = NULL
)
}
\arguments{
\item{Name}{[required] The name of the connection to update.}

\item{Description}{A description for the connection.}

\item{AuthorizationType}{The type of authorization to use for the connection.}

\item{AuthParameters}{The authorization parameters to use for the connection.}

\item{InvocationConnectivityParameters}{For connections to private APIs, the parameters to use for invoking the
API.

For more information, see \href{https://docs.aws.amazon.com/eventbridge/latest/userguide/connection-private.html}{Connecting to private APIs}
in the \emph{Amazon EventBridge User Guide} .}

\item{KmsKeyIdentifier}{The identifier of the KMS customer managed key for EventBridge to use,
if you choose to use a customer managed key to encrypt this connection.
The identifier can be the key Amazon Resource Name (ARN), KeyId, key
alias, or key alias ARN.

If you do not specify a customer managed key identifier, EventBridge
uses an Amazon Web Services owned key to encrypt the connection.

For more information, see \href{https://docs.aws.amazon.com/kms/latest/developerguide/viewing-keys.html}{Identify and view keys}
in the \emph{Key Management Service Developer Guide}.}
}
\description{
Updates settings for a connection.

See \url{https://www.paws-r-sdk.com/docs/eventbridge_update_connection/} for full documentation.
}
\keyword{internal}
