% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridgepipes_operations.R
\name{eventbridgepipes_create_pipe}
\alias{eventbridgepipes_create_pipe}
\title{Create a pipe}
\usage{
eventbridgepipes_create_pipe(
  Name,
  Description = NULL,
  DesiredState = NULL,
  Source,
  SourceParameters = NULL,
  Enrichment = NULL,
  EnrichmentParameters = NULL,
  Target,
  TargetParameters = NULL,
  RoleArn,
  Tags = NULL,
  LogConfiguration = NULL,
  KmsKeyIdentifier = NULL
)
}
\arguments{
\item{Name}{[required] The name of the pipe.}

\item{Description}{A description of the pipe.}

\item{DesiredState}{The state the pipe should be in.}

\item{Source}{[required] The ARN of the source resource.}

\item{SourceParameters}{The parameters required to set up a source for your pipe.}

\item{Enrichment}{The ARN of the enrichment resource.}

\item{EnrichmentParameters}{The parameters required to set up enrichment on your pipe.}

\item{Target}{[required] The ARN of the target resource.}

\item{TargetParameters}{The parameters required to set up a target for your pipe.

For more information about pipe target parameters, including how to use
dynamic path parameters, see \href{https://docs.aws.amazon.com/eventbridge/latest/userguide/eb-pipes-event-target.html}{Target parameters}
in the \emph{Amazon EventBridge User Guide}.}

\item{RoleArn}{[required] The ARN of the role that allows the pipe to send data to the target.}

\item{Tags}{The list of key-value pairs to associate with the pipe.}

\item{LogConfiguration}{The logging configuration settings for the pipe.}

\item{KmsKeyIdentifier}{The identifier of the KMS customer managed key for EventBridge to use,
if you choose to use a customer managed key to encrypt pipe data. The
identifier can be the key Amazon Resource Name (ARN), KeyId, key alias,
or key alias ARN.

If you do not specify a customer managed key identifier, EventBridge
uses an Amazon Web Services owned key to encrypt pipe data.

For more information, see \href{https://docs.aws.amazon.com/kms/latest/developerguide/}{Managing keys} in the
\emph{Key Management Service Developer Guide}.}
}
\description{
Create a pipe. Amazon EventBridge Pipes connect event sources to targets and reduces the need for specialized knowledge and integration code.

See \url{https://www.paws-r-sdk.com/docs/eventbridgepipes_create_pipe/} for full documentation.
}
\keyword{internal}
