% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridgepipes_operations.R
\name{eventbridgepipes_delete_pipe}
\alias{eventbridgepipes_delete_pipe}
\title{Delete an existing pipe}
\usage{
eventbridgepipes_delete_pipe(Name)
}
\arguments{
\item{Name}{[required] The name of the pipe.}
}
\description{
Delete an existing pipe. For more information about pipes, see \href{https://docs.aws.amazon.com/eventbridge/latest/userguide/eb-pipes.html}{Amazon EventBridge Pipes} in the Amazon EventBridge User Guide.

See \url{https://www.paws-r-sdk.com/docs/eventbridgepipes_delete_pipe/} for full documentation.
}
\keyword{internal}
