% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridgepipes_operations.R
\name{eventbridgepipes_list_tags_for_resource}
\alias{eventbridgepipes_list_tags_for_resource}
\title{Displays the tags associated with a pipe}
\usage{
eventbridgepipes_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the pipe for which you want to view tags.}
}
\description{
Displays the tags associated with a pipe.

See \url{https://www.paws-r-sdk.com/docs/eventbridgepipes_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
