% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridgepipes_operations.R
\name{eventbridgepipes_stop_pipe}
\alias{eventbridgepipes_stop_pipe}
\title{Stop an existing pipe}
\usage{
eventbridgepipes_stop_pipe(Name)
}
\arguments{
\item{Name}{[required] The name of the pipe.}
}
\description{
Stop an existing pipe.

See \url{https://www.paws-r-sdk.com/docs/eventbridgepipes_stop_pipe/} for full documentation.
}
\keyword{internal}
