% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridgepipes_operations.R
\name{eventbridgepipes_tag_resource}
\alias{eventbridgepipes_tag_resource}
\title{Assigns one or more tags (key-value pairs) to the specified pipe}
\usage{
eventbridgepipes_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The ARN of the pipe.}

\item{tags}{[required] The list of key-value pairs associated with the pipe.}
}
\description{
Assigns one or more tags (key-value pairs) to the specified pipe. Tags can help you organize and categorize your resources. You can also use them to scope user permissions by granting a user permission to access or change only resources with certain tag values.

See \url{https://www.paws-r-sdk.com/docs/eventbridgepipes_tag_resource/} for full documentation.
}
\keyword{internal}
