% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridgepipes_operations.R
\name{eventbridgepipes_untag_resource}
\alias{eventbridgepipes_untag_resource}
\title{Removes one or more tags from the specified pipes}
\usage{
eventbridgepipes_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The ARN of the pipe.}

\item{tagKeys}{[required] The list of tag keys to remove from the pipe.}
}
\description{
Removes one or more tags from the specified pipes.

See \url{https://www.paws-r-sdk.com/docs/eventbridgepipes_untag_resource/} for full documentation.
}
\keyword{internal}
