% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridgepipes_operations.R
\name{eventbridgepipes_update_pipe}
\alias{eventbridgepipes_update_pipe}
\title{Update an existing pipe}
\usage{
eventbridgepipes_update_pipe(
  Description = NULL,
  DesiredState = NULL,
  Enrichment = NULL,
  EnrichmentParameters = NULL,
  Name,
  RoleArn,
  SourceParameters = NULL,
  Target = NULL,
  TargetParameters = NULL
)
}
\arguments{
\item{Description}{A description of the pipe.}

\item{DesiredState}{The state the pipe should be in.}

\item{Enrichment}{The ARN of the enrichment resource.}

\item{EnrichmentParameters}{The parameters required to set up enrichment on your pipe.}

\item{Name}{[required] The name of the pipe.}

\item{RoleArn}{[required] The ARN of the role that allows the pipe to send data to the target.}

\item{SourceParameters}{The parameters required to set up a source for your pipe.}

\item{Target}{The ARN of the target resource.}

\item{TargetParameters}{The parameters required to set up a target for your pipe.}
}
\description{
Update an existing pipe. When you call \code{\link[=eventbridgepipes_update_pipe]{update_pipe}}, only the fields that are included in the request are changed, the rest are unchanged. The exception to this is if you modify any Amazon Web Services-service specific fields in the \code{SourceParameters}, \code{EnrichmentParameters}, or \code{TargetParameters} objects. The fields in these objects are updated atomically as one and override existing values. This is by design and means that if you don't specify an optional field in one of these Parameters objects, that field will be set to its system-default value after the update.

See \url{https://www.paws-r-sdk.com/docs/eventbridgepipes_update_pipe/} for full documentation.
}
\keyword{internal}
