% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridgescheduler_operations.R
\name{eventbridgescheduler_create_schedule_group}
\alias{eventbridgescheduler_create_schedule_group}
\title{Creates the specified schedule group}
\usage{
eventbridgescheduler_create_schedule_group(
  ClientToken = NULL,
  Name,
  Tags = NULL
)
}
\arguments{
\item{ClientToken}{Unique, case-sensitive identifier you provide to ensure the idempotency
of the request. If you do not specify a client token, EventBridge
Scheduler uses a randomly generated token for the request to ensure
idempotency.}

\item{Name}{[required] The name of the schedule group that you are creating.}

\item{Tags}{The list of tags to associate with the schedule group.}
}
\description{
Creates the specified schedule group.

See \url{https://www.paws-r-sdk.com/docs/eventbridgescheduler_create_schedule_group/} for full documentation.
}
\keyword{internal}
