% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridgescheduler_operations.R
\name{eventbridgescheduler_delete_schedule}
\alias{eventbridgescheduler_delete_schedule}
\title{Deletes the specified schedule}
\usage{
eventbridgescheduler_delete_schedule(
  ClientToken = NULL,
  GroupName = NULL,
  Name
)
}
\arguments{
\item{ClientToken}{Unique, case-sensitive identifier you provide to ensure the idempotency
of the request. If you do not specify a client token, EventBridge
Scheduler uses a randomly generated token for the request to ensure
idempotency.}

\item{GroupName}{The name of the schedule group associated with this schedule. If you
omit this, the default schedule group is used.}

\item{Name}{[required] The name of the schedule to delete.}
}
\description{
Deletes the specified schedule.

See \url{https://www.paws-r-sdk.com/docs/eventbridgescheduler_delete_schedule/} for full documentation.
}
\keyword{internal}
