% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridgescheduler_operations.R
\name{eventbridgescheduler_get_schedule}
\alias{eventbridgescheduler_get_schedule}
\title{Retrieves the specified schedule}
\usage{
eventbridgescheduler_get_schedule(GroupName = NULL, Name)
}
\arguments{
\item{GroupName}{The name of the schedule group associated with this schedule. If you
omit this, EventBridge Scheduler assumes that the schedule is associated
with the default group.}

\item{Name}{[required] The name of the schedule to retrieve.}
}
\description{
Retrieves the specified schedule.

See \url{https://www.paws-r-sdk.com/docs/eventbridgescheduler_get_schedule/} for full documentation.
}
\keyword{internal}
