% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridgescheduler_operations.R
\name{eventbridgescheduler_get_schedule_group}
\alias{eventbridgescheduler_get_schedule_group}
\title{Retrieves the specified schedule group}
\usage{
eventbridgescheduler_get_schedule_group(Name)
}
\arguments{
\item{Name}{[required] The name of the schedule group to retrieve.}
}
\description{
Retrieves the specified schedule group.

See \url{https://www.paws-r-sdk.com/docs/eventbridgescheduler_get_schedule_group/} for full documentation.
}
\keyword{internal}
