% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridgescheduler_operations.R
\name{eventbridgescheduler_list_schedule_groups}
\alias{eventbridgescheduler_list_schedule_groups}
\title{Returns a paginated list of your schedule groups}
\usage{
eventbridgescheduler_list_schedule_groups(
  MaxResults = NULL,
  NamePrefix = NULL,
  NextToken = NULL
)
}
\arguments{
\item{MaxResults}{If specified, limits the number of results returned by this operation.
The operation also returns a \code{NextToken} which you can use in a
subsequent operation to retrieve the next set of results.}

\item{NamePrefix}{The name prefix that you can use to return a filtered list of your
schedule groups.}

\item{NextToken}{The token returned by a previous call to retrieve the next set of
results.}
}
\description{
Returns a paginated list of your schedule groups.

See \url{https://www.paws-r-sdk.com/docs/eventbridgescheduler_list_schedule_groups/} for full documentation.
}
\keyword{internal}
