% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridgescheduler_operations.R
\name{eventbridgescheduler_list_schedules}
\alias{eventbridgescheduler_list_schedules}
\title{Returns a paginated list of your EventBridge Scheduler schedules}
\usage{
eventbridgescheduler_list_schedules(
  GroupName = NULL,
  MaxResults = NULL,
  NamePrefix = NULL,
  NextToken = NULL,
  State = NULL
)
}
\arguments{
\item{GroupName}{If specified, only lists the schedules whose associated schedule group
matches the given filter.}

\item{MaxResults}{If specified, limits the number of results returned by this operation.
The operation also returns a \code{NextToken} which you can use in a
subsequent operation to retrieve the next set of results.}

\item{NamePrefix}{Schedule name prefix to return the filtered list of resources.}

\item{NextToken}{The token returned by a previous call to retrieve the next set of
results.}

\item{State}{If specified, only lists the schedules whose current state matches the
given filter.}
}
\description{
Returns a paginated list of your EventBridge Scheduler schedules.

See \url{https://www.paws-r-sdk.com/docs/eventbridgescheduler_list_schedules/} for full documentation.
}
\keyword{internal}
