% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridgescheduler_operations.R
\name{eventbridgescheduler_list_tags_for_resource}
\alias{eventbridgescheduler_list_tags_for_resource}
\title{Lists the tags associated with the Scheduler resource}
\usage{
eventbridgescheduler_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the EventBridge Scheduler resource for which you want to view
tags.}
}
\description{
Lists the tags associated with the Scheduler resource.

See \url{https://www.paws-r-sdk.com/docs/eventbridgescheduler_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
