% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridgescheduler_operations.R
\name{eventbridgescheduler_tag_resource}
\alias{eventbridgescheduler_tag_resource}
\title{Assigns one or more tags (key-value pairs) to the specified EventBridge
Scheduler resource}
\usage{
eventbridgescheduler_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the schedule group that you are adding
tags to.}

\item{Tags}{[required] The list of tags to associate with the schedule group.}
}
\description{
Assigns one or more tags (key-value pairs) to the specified EventBridge Scheduler resource. You can only assign tags to schedule groups.

See \url{https://www.paws-r-sdk.com/docs/eventbridgescheduler_tag_resource/} for full documentation.
}
\keyword{internal}
