% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridgescheduler_operations.R
\name{eventbridgescheduler_untag_resource}
\alias{eventbridgescheduler_untag_resource}
\title{Removes one or more tags from the specified EventBridge Scheduler
schedule group}
\usage{
eventbridgescheduler_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the schedule group from which you are
removing tags.}

\item{TagKeys}{[required] The list of tag keys to remove from the resource.}
}
\description{
Removes one or more tags from the specified EventBridge Scheduler schedule group.

See \url{https://www.paws-r-sdk.com/docs/eventbridgescheduler_untag_resource/} for full documentation.
}
\keyword{internal}
