% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_batch_delete_device_position_history}
\alias{locationservice_batch_delete_device_position_history}
\title{Deletes the position history of one or more devices from a tracker
resource}
\usage{
locationservice_batch_delete_device_position_history(TrackerName, DeviceIds)
}
\arguments{
\item{TrackerName}{[required] The name of the tracker resource to delete the device position history
from.}

\item{DeviceIds}{[required] Devices whose position history you want to delete.
\itemize{
\item For example, for two devices: \verb{“DeviceIds” : [DeviceId1,DeviceId2]}
}}
}
\description{
Deletes the position history of one or more devices from a tracker resource.

See \url{https://www.paws-r-sdk.com/docs/locationservice_batch_delete_device_position_history/} for full documentation.
}
\keyword{internal}
