% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_batch_delete_geofence}
\alias{locationservice_batch_delete_geofence}
\title{Deletes a batch of geofences from a geofence collection}
\usage{
locationservice_batch_delete_geofence(CollectionName, GeofenceIds)
}
\arguments{
\item{CollectionName}{[required] The geofence collection storing the geofences to be deleted.}

\item{GeofenceIds}{[required] The batch of geofences to be deleted.}
}
\description{
Deletes a batch of geofences from a geofence collection.

See \url{https://paws-r.github.io/docs/locationservice/batch_delete_geofence.html} for full documentation.
}
\keyword{internal}
