% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_batch_evaluate_geofences}
\alias{locationservice_batch_evaluate_geofences}
\title{Evaluates device positions against the geofence geometries from a given
geofence collection}
\usage{
locationservice_batch_evaluate_geofences(CollectionName, DevicePositionUpdates)
}
\arguments{
\item{CollectionName}{[required] The geofence collection used in evaluating the position of devices
against its geofences.}

\item{DevicePositionUpdates}{[required] Contains device details for each device to be evaluated against the
given geofence collection.}
}
\description{
Evaluates device positions against the geofence geometries from a given geofence collection.

See \url{https://www.paws-r-sdk.com/docs/locationservice_batch_evaluate_geofences/} for full documentation.
}
\keyword{internal}
