% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_batch_get_device_position}
\alias{locationservice_batch_get_device_position}
\title{Lists the latest device positions for requested devices}
\usage{
locationservice_batch_get_device_position(TrackerName, DeviceIds)
}
\arguments{
\item{TrackerName}{[required] The tracker resource retrieving the device position.}

\item{DeviceIds}{[required] Devices whose position you want to retrieve.
\itemize{
\item For example, for two devices:
\code{device-ids=DeviceId1&device-ids=DeviceId2}
}}
}
\description{
Lists the latest device positions for requested devices.

See \url{https://www.paws-r-sdk.com/docs/locationservice_batch_get_device_position/} for full documentation.
}
\keyword{internal}
