% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_batch_put_geofence}
\alias{locationservice_batch_put_geofence}
\title{A batch request for storing geofence geometries into a given geofence
collection, or updates the geometry of an existing geofence if a
geofence ID is included in the request}
\usage{
locationservice_batch_put_geofence(CollectionName, Entries)
}
\arguments{
\item{CollectionName}{[required] The geofence collection storing the geofences.}

\item{Entries}{[required] The batch of geofences to be stored in a geofence collection.}
}
\description{
A batch request for storing geofence geometries into a given geofence collection, or updates the geometry of an existing geofence if a geofence ID is included in the request.

See \url{https://www.paws-r-sdk.com/docs/locationservice_batch_put_geofence/} for full documentation.
}
\keyword{internal}
