% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_batch_update_device_position}
\alias{locationservice_batch_update_device_position}
\title{Uploads position update data for one or more devices to a tracker
resource}
\usage{
locationservice_batch_update_device_position(TrackerName, Updates)
}
\arguments{
\item{TrackerName}{[required] The name of the tracker resource to update.}

\item{Updates}{[required] Contains the position update details for each device.}
}
\description{
Uploads position update data for one or more devices to a tracker resource. Amazon Location uses the data when it reports the last known device position and position history. Amazon Location retains location data for 30 days.

See \url{https://www.paws-r-sdk.com/docs/locationservice_batch_update_device_position/} for full documentation.
}
\keyword{internal}
