% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_calculate_route_matrix}
\alias{locationservice_calculate_route_matrix}
\title{This operation is no longer current and may be deprecated in the future}
\usage{
locationservice_calculate_route_matrix(
  CalculatorName,
  DeparturePositions,
  DestinationPositions,
  TravelMode = NULL,
  DepartureTime = NULL,
  DepartNow = NULL,
  DistanceUnit = NULL,
  CarModeOptions = NULL,
  TruckModeOptions = NULL,
  Key = NULL
)
}
\arguments{
\item{CalculatorName}{[required] The name of the route calculator resource that you want to use to calculate the route matrix.}

\item{DeparturePositions}{[required] The list of departure (origin) positions for the route matrix. An array of points, each of which is itself a 2-value array defined in \href{https://earth-info.nga.mil/index.php?dir=wgs84&action=wgs84}{WGS 84} format: \verb{[longitude, latitude]}. For example, \verb{[-123.115, 49.285]}.

Depending on the data provider selected in the route calculator resource there may be additional restrictions on the inputs you can choose. See \href{https://docs.aws.amazon.com/location/previous/developerguide/calculate-route-matrix.html#matrix-routing-position-limits}{Position restrictions} in the \emph{Amazon Location Service Developer Guide}.

For route calculators that use Esri as the data provider, if you specify a departure that's not located on a road, Amazon Location \href{https://docs.aws.amazon.com/location/previous/developerguide/snap-to-nearby-road.html}{moves the position to the nearest road}. The snapped value is available in the result in \code{SnappedDeparturePositions}.

Valid Values: \verb{[-180 to 180,-90 to 90]}}

\item{DestinationPositions}{[required] The list of destination positions for the route matrix. An array of points, each of which is itself a 2-value array defined in \href{https://earth-info.nga.mil/index.php?dir=wgs84&action=wgs84}{WGS 84} format: \verb{[longitude, latitude]}. For example, \verb{[-122.339, 47.615]}

Depending on the data provider selected in the route calculator resource there may be additional restrictions on the inputs you can choose. See \href{https://docs.aws.amazon.com/location/previous/developerguide/calculate-route-matrix.html#matrix-routing-position-limits}{Position restrictions} in the \emph{Amazon Location Service Developer Guide}.

For route calculators that use Esri as the data provider, if you specify a destination that's not located on a road, Amazon Location \href{https://docs.aws.amazon.com/location/previous/developerguide/snap-to-nearby-road.html}{moves the position to the nearest road}. The snapped value is available in the result in \code{SnappedDestinationPositions}.

Valid Values: \verb{[-180 to 180,-90 to 90]}}

\item{TravelMode}{Specifies the mode of transport when calculating a route. Used in estimating the speed of travel and road compatibility.

The \code{TravelMode} you specify also determines how you specify route preferences:
\itemize{
\item If traveling by \code{Car} use the \code{CarModeOptions} parameter.
\item If traveling by \code{Truck} use the \code{TruckModeOptions} parameter.
}

\code{Bicycle} or \code{Motorcycle} are only valid when using \code{Grab} as a data provider, and only within Southeast Asia.

\code{Truck} is not available for Grab.

For more information about using Grab as a data provider, see \href{https://docs.aws.amazon.com/location/previous/developerguide/grab.html}{GrabMaps} in the \emph{Amazon Location Service Developer Guide}.

Default Value: \code{Car}}

\item{DepartureTime}{Specifies the desired time of departure. Uses the given time to calculate the route matrix. You can't set both \code{DepartureTime} and \code{DepartNow}. If neither is set, the best time of day to travel with the best traffic conditions is used to calculate the route matrix.

Setting a departure time in the past returns a \verb{400 ValidationException} error.
\itemize{
\item In ISO 8601 format: \code{YYYY-MM-DDThh:mm:ss.sssZ}. For example, \verb{2020–07-2T12:15:20.000Z+01:00}
}}

\item{DepartNow}{Sets the time of departure as the current time. Uses the current time to calculate the route matrix. You can't set both \code{DepartureTime} and \code{DepartNow}. If neither is set, the best time of day to travel with the best traffic conditions is used to calculate the route matrix.

Default Value: \code{false}

Valid Values: \code{false} | \code{true}}

\item{DistanceUnit}{Set the unit system to specify the distance.

Default Value: \code{Kilometers}}

\item{CarModeOptions}{Specifies route preferences when traveling by \code{Car}, such as avoiding routes that use ferries or tolls.

Requirements: \code{TravelMode} must be specified as \code{Car}.}

\item{TruckModeOptions}{Specifies route preferences when traveling by \code{Truck}, such as avoiding routes that use ferries or tolls, and truck specifications to consider when choosing an optimal road.

Requirements: \code{TravelMode} must be specified as \code{Truck}.}

\item{Key}{The optional \href{https://docs.aws.amazon.com/location/previous/developerguide/using-apikeys.html}{API key} to authorize the request.}
}
\description{
This operation is no longer current and may be deprecated in the future. We recommend you upgrade to the V2 \code{calculate_route_matrix} unless you require Grab data.

See \url{https://www.paws-r-sdk.com/docs/locationservice_calculate_route_matrix/} for full documentation.
}
\keyword{internal}
