% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_delete_geofence_collection}
\alias{locationservice_delete_geofence_collection}
\title{Deletes a geofence collection from your Amazon Web Services account}
\usage{
locationservice_delete_geofence_collection(CollectionName)
}
\arguments{
\item{CollectionName}{[required] The name of the geofence collection to be deleted.}
}
\description{
Deletes a geofence collection from your Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/locationservice_delete_geofence_collection/} for full documentation.
}
\keyword{internal}
