% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_delete_key}
\alias{locationservice_delete_key}
\title{Deletes the specified API key}
\usage{
locationservice_delete_key(KeyName)
}
\arguments{
\item{KeyName}{[required] The name of the API key to delete.}
}
\description{
Deletes the specified API key. The API key must have been deactivated more than 90 days previously.

See \url{https://www.paws-r-sdk.com/docs/locationservice_delete_key/} for full documentation.
}
\keyword{internal}
