% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_delete_map}
\alias{locationservice_delete_map}
\title{Deletes a map resource from your Amazon Web Services account}
\usage{
locationservice_delete_map(MapName)
}
\arguments{
\item{MapName}{[required] The name of the map resource to be deleted.}
}
\description{
Deletes a map resource from your Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/locationservice_delete_map/} for full documentation.
}
\keyword{internal}
