% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_delete_place_index}
\alias{locationservice_delete_place_index}
\title{Deletes a place index resource from your Amazon Web Services account}
\usage{
locationservice_delete_place_index(IndexName)
}
\arguments{
\item{IndexName}{[required] The name of the place index resource to be deleted.}
}
\description{
Deletes a place index resource from your Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/locationservice_delete_place_index/} for full documentation.
}
\keyword{internal}
