% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_delete_route_calculator}
\alias{locationservice_delete_route_calculator}
\title{Deletes a route calculator resource from your AWS account}
\usage{
locationservice_delete_route_calculator(CalculatorName)
}
\arguments{
\item{CalculatorName}{[required] The name of the route calculator resource to be deleted.}
}
\description{
Deletes a route calculator resource from your AWS account.

See \url{https://paws-r.github.io/docs/locationservice/delete_route_calculator.html} for full documentation.
}
\keyword{internal}
