% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_delete_tracker}
\alias{locationservice_delete_tracker}
\title{Deletes a tracker resource from your AWS account}
\usage{
locationservice_delete_tracker(TrackerName)
}
\arguments{
\item{TrackerName}{[required] The name of the tracker resource to be deleted.}
}
\description{
Deletes a tracker resource from your AWS account.

See \url{https://paws-r.github.io/docs/locationservice/delete_tracker.html} for full documentation.
}
\keyword{internal}
