% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_describe_geofence_collection}
\alias{locationservice_describe_geofence_collection}
\title{Retrieves the geofence collection details}
\usage{
locationservice_describe_geofence_collection(CollectionName)
}
\arguments{
\item{CollectionName}{[required] The name of the geofence collection.}
}
\description{
Retrieves the geofence collection details.

See \url{https://www.paws-r-sdk.com/docs/locationservice_describe_geofence_collection/} for full documentation.
}
\keyword{internal}
