% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_describe_map}
\alias{locationservice_describe_map}
\title{Retrieves the map resource details}
\usage{
locationservice_describe_map(MapName)
}
\arguments{
\item{MapName}{[required] The name of the map resource.}
}
\description{
Retrieves the map resource details.

See \url{https://paws-r.github.io/docs/locationservice/describe_map.html} for full documentation.
}
\keyword{internal}
