% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_describe_place_index}
\alias{locationservice_describe_place_index}
\title{Retrieves the place index resource details}
\usage{
locationservice_describe_place_index(IndexName)
}
\arguments{
\item{IndexName}{[required] The name of the place index resource.}
}
\description{
Retrieves the place index resource details.

See \url{https://paws-r.github.io/docs/locationservice/describe_place_index.html} for full documentation.
}
\keyword{internal}
