% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_describe_tracker}
\alias{locationservice_describe_tracker}
\title{Retrieves the tracker resource details}
\usage{
locationservice_describe_tracker(TrackerName)
}
\arguments{
\item{TrackerName}{[required] The name of the tracker resource.}
}
\description{
Retrieves the tracker resource details.

See \url{https://paws-r.github.io/docs/locationservice/describe_tracker.html} for full documentation.
}
\keyword{internal}
