% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_disassociate_tracker_consumer}
\alias{locationservice_disassociate_tracker_consumer}
\title{Removes the association between a tracker resource and a geofence
collection}
\usage{
locationservice_disassociate_tracker_consumer(TrackerName, ConsumerArn)
}
\arguments{
\item{TrackerName}{[required] The name of the tracker resource to be dissociated from the consumer.}

\item{ConsumerArn}{[required] The Amazon Resource Name (ARN) for the geofence collection to be
disassociated from the tracker resource. Used when you need to specify a
resource across all Amazon Web Services.
\itemize{
\item Format example:
\code{arn:aws:geo:region:account-id:geofence-collection/ExampleGeofenceCollectionConsumer}
}}
}
\description{
Removes the association between a tracker resource and a geofence collection.

See \url{https://www.paws-r-sdk.com/docs/locationservice_disassociate_tracker_consumer/} for full documentation.
}
\keyword{internal}
