% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_get_geofence}
\alias{locationservice_get_geofence}
\title{Retrieves the geofence details from a geofence collection}
\usage{
locationservice_get_geofence(CollectionName, GeofenceId)
}
\arguments{
\item{CollectionName}{[required] The geofence collection storing the target geofence.}

\item{GeofenceId}{[required] The geofence you're retrieving details for.}
}
\description{
Retrieves the geofence details from a geofence collection.

See \url{https://www.paws-r-sdk.com/docs/locationservice_get_geofence/} for full documentation.
}
\keyword{internal}
