% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_get_job}
\alias{locationservice_get_job}
\title{GetJob retrieves detailed information about a specific job, including
its current status, configuration, and error information if the job
failed}
\usage{
locationservice_get_job(JobId)
}
\arguments{
\item{JobId}{[required] The unique identifier of the job to retrieve.}
}
\description{
\code{\link[=locationservice_get_job]{get_job}} retrieves detailed information about a specific job, including its current status, configuration, and error information if the job failed.

See \url{https://www.paws-r-sdk.com/docs/locationservice_get_job/} for full documentation.
}
\keyword{internal}
