% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_get_map_sprites}
\alias{locationservice_get_map_sprites}
\title{Retrieves the sprite sheet corresponding to a map resource}
\usage{
locationservice_get_map_sprites(FileName, MapName)
}
\arguments{
\item{FileName}{[required] The name of the sprite ﬁle. Use the following ﬁle names for the sprite
sheet:
\itemize{
\item \code{sprites.png}
\item \verb{sprites@2x.png} for high pixel density displays
}

For the JSON document contain image offsets. Use the following ﬁle
names:
\itemize{
\item \code{sprites.json}
\item \verb{sprites@2x.json} for high pixel density displays
}}

\item{MapName}{[required] The map resource associated with the sprite ﬁle.}
}
\description{
Retrieves the sprite sheet corresponding to a map resource. The sprite sheet is a PNG image paired with a JSON document describing the offsets of individual icons that will be displayed on a rendered map.

See \url{https://paws-r.github.io/docs/locationservice/get_map_sprites.html} for full documentation.
}
\keyword{internal}
