% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_get_map_style_descriptor}
\alias{locationservice_get_map_style_descriptor}
\title{Retrieves the map style descriptor from a map resource}
\usage{
locationservice_get_map_style_descriptor(MapName)
}
\arguments{
\item{MapName}{[required] The map resource to retrieve the style descriptor from.}
}
\description{
Retrieves the map style descriptor from a map resource.

See \url{https://paws-r.github.io/docs/locationservice/get_map_style_descriptor.html} for full documentation.
}
\keyword{internal}
