% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_list_device_positions}
\alias{locationservice_list_device_positions}
\title{A batch request to retrieve all device positions}
\usage{
locationservice_list_device_positions(
  MaxResults = NULL,
  NextToken = NULL,
  TrackerName
)
}
\arguments{
\item{MaxResults}{An optional limit for the number of entries returned in a single call.

Default value: \code{100}}

\item{NextToken}{The pagination token specifying which page of results to return in the
response. If no token is provided, the default page is the first page.

Default value: \code{null}}

\item{TrackerName}{[required] The tracker resource containing the requested devices.}
}
\description{
A batch request to retrieve all device positions.

See \url{https://paws-r.github.io/docs/locationservice/list_device_positions.html} for full documentation.
}
\keyword{internal}
