% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_list_jobs}
\alias{locationservice_list_jobs}
\title{ListJobs retrieves a list of jobs with optional filtering and pagination
support}
\usage{
locationservice_list_jobs(Filter = NULL, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{Filter}{An optional structure containing criteria by which to filter job
results.}

\item{MaxResults}{Maximum number of jobs to return.}

\item{NextToken}{The pagination token specifying which page of results to return in the
response. If no token is provided, the default page is the first page.}
}
\description{
\code{\link[=locationservice_list_jobs]{list_jobs}} retrieves a list of jobs with optional filtering and pagination support.

See \url{https://www.paws-r-sdk.com/docs/locationservice_list_jobs/} for full documentation.
}
\keyword{internal}
