% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_list_maps}
\alias{locationservice_list_maps}
\title{This operation is no longer current and may be deprecated in the future}
\usage{
locationservice_list_maps(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{An optional limit for the number of resources returned in a single call.

Default value: \code{100}}

\item{NextToken}{The pagination token specifying which page of results to return in the
response. If no token is provided, the default page is the first page.

Default value: \code{null}}
}
\description{
This operation is no longer current and may be deprecated in the future. We recommend upgrading to the Maps API V2 unless you require \code{Grab} data.

See \url{https://www.paws-r-sdk.com/docs/locationservice_list_maps/} for full documentation.
}
\keyword{internal}
