% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_list_tags_for_resource}
\alias{locationservice_list_tags_for_resource}
\title{Returns a list of tags that are applied to the specified Amazon Location
resource}
\usage{
locationservice_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource whose tags you want to
retrieve.
\itemize{
\item Format example:
\code{arn:aws:geo:region:account-id:resourcetype/ExampleResource}
}}
}
\description{
Returns a list of tags that are applied to the specified Amazon Location resource.

See \url{https://www.paws-r-sdk.com/docs/locationservice_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
