% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_list_tracker_consumers}
\alias{locationservice_list_tracker_consumers}
\title{Lists geofence collections currently associated to the given tracker
resource}
\usage{
locationservice_list_tracker_consumers(
  MaxResults = NULL,
  NextToken = NULL,
  TrackerName
)
}
\arguments{
\item{MaxResults}{An optional limit for the number of resources returned in a single call.

Default value: \code{100}}

\item{NextToken}{The pagination token specifying which page of results to return in the
response. If no token is provided, the default page is the first page.

Default value: \code{null}}

\item{TrackerName}{[required] The tracker resource whose associated geofence collections you want to
list.}
}
\description{
Lists geofence collections currently associated to the given tracker resource.

See \url{https://paws-r.github.io/docs/locationservice/list_tracker_consumers.html} for full documentation.
}
\keyword{internal}
