% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_start_job}
\alias{locationservice_start_job}
\title{StartJob starts a new asynchronous bulk processing job}
\usage{
locationservice_start_job(
  ClientToken = NULL,
  Action,
  ActionOptions = NULL,
  ExecutionRoleArn,
  InputOptions,
  Name = NULL,
  OutputOptions,
  Tags = NULL
)
}
\arguments{
\item{ClientToken}{A unique identifier for this request to ensure idempotency.}

\item{Action}{[required] The action to perform on the input data.}

\item{ActionOptions}{Additional parameters that can be requested for each result.}

\item{ExecutionRoleArn}{[required] The Amazon Resource Name (ARN) of the IAM role that Amazon Location
Service assumes during job processing. Amazon Location Service uses this
role to access the input and output locations specified for the job.

The IAM role must be created in the same Amazon Web Services account
where you plan to run your job.

For more information about configuring IAM roles for Amazon Location
jobs, see \href{https://docs.aws.amazon.com/location/latest/developerguide/configure-iam-role-policy-credentials.html}{Configure IAM permissions}
in the \emph{Amazon Location Service Developer Guide}.}

\item{InputOptions}{[required] Configuration for input data location and format.

Input files have a limitation of 10gb per file, and 1gb per Parquet
row-group within the file.}

\item{Name}{An optional name for the job resource.}

\item{OutputOptions}{[required] Configuration for output data location and format.}

\item{Tags}{Tags and corresponding values to be associated with the job.}
}
\description{
\code{\link[=locationservice_start_job]{start_job}} starts a new asynchronous bulk processing job. You specify the input data location in Amazon S3, the action to perform, and the output location where results are written.

See \url{https://www.paws-r-sdk.com/docs/locationservice_start_job/} for full documentation.
}
\keyword{internal}
