% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_untag_resource}
\alias{locationservice_untag_resource}
\title{Removes one or more tags from the specified Amazon Location resource}
\usage{
locationservice_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource from which you want to
remove tags.
\itemize{
\item Format example:
\code{arn:aws:geo:region:account-id:resourcetype/ExampleResource}
}}

\item{TagKeys}{[required] The list of tag keys to remove from the specified resource.}
}
\description{
Removes one or more tags from the specified Amazon Location resource.

See \url{https://www.paws-r-sdk.com/docs/locationservice_untag_resource/} for full documentation.
}
\keyword{internal}
