% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_update_geofence_collection}
\alias{locationservice_update_geofence_collection}
\title{Updates the specified properties of a given geofence collection}
\usage{
locationservice_update_geofence_collection(
  CollectionName,
  Description = NULL,
  PricingPlan = NULL,
  PricingPlanDataSource = NULL
)
}
\arguments{
\item{CollectionName}{[required] The name of the geofence collection to update.}

\item{Description}{Updates the description for the geofence collection.}

\item{PricingPlan}{No longer used. If included, the only allowed value is
\code{RequestBasedUsage}.}

\item{PricingPlanDataSource}{This parameter is no longer used.}
}
\description{
Updates the specified properties of a given geofence collection.

See \url{https://paws-r.github.io/docs/locationservice/update_geofence_collection.html} for full documentation.
}
\keyword{internal}
