% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_update_map}
\alias{locationservice_update_map}
\title{Updates the specified properties of a given map resource}
\usage{
locationservice_update_map(Description = NULL, MapName, PricingPlan = NULL)
}
\arguments{
\item{Description}{Updates the description for the map resource.}

\item{MapName}{[required] The name of the map resource to update.}

\item{PricingPlan}{No longer used. If included, the only allowed value is
\code{RequestBasedUsage}.}
}
\description{
Updates the specified properties of a given map resource.

See \url{https://paws-r.github.io/docs/locationservice/update_map.html} for full documentation.
}
\keyword{internal}
