% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_update_place_index}
\alias{locationservice_update_place_index}
\title{Updates the specified properties of a given place index resource}
\usage{
locationservice_update_place_index(
  IndexName,
  PricingPlan = NULL,
  Description = NULL,
  DataSourceConfiguration = NULL
)
}
\arguments{
\item{IndexName}{[required] The name of the place index resource to update.}

\item{PricingPlan}{No longer used. If included, the only allowed value is
\code{RequestBasedUsage}.}

\item{Description}{Updates the description for the place index resource.}

\item{DataSourceConfiguration}{Updates the data storage option for the place index resource.}
}
\description{
Updates the specified properties of a given place index resource.

See \url{https://www.paws-r-sdk.com/docs/locationservice_update_place_index/} for full documentation.
}
\keyword{internal}
