% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_update_route_calculator}
\alias{locationservice_update_route_calculator}
\title{Updates the specified properties for a given route calculator resource}
\usage{
locationservice_update_route_calculator(
  CalculatorName,
  Description = NULL,
  PricingPlan = NULL
)
}
\arguments{
\item{CalculatorName}{[required] The name of the route calculator resource to update.}

\item{Description}{Updates the description for the route calculator resource.}

\item{PricingPlan}{No longer used. If included, the only allowed value is
\code{RequestBasedUsage}.}
}
\description{
Updates the specified properties for a given route calculator resource.

See \url{https://paws-r.github.io/docs/locationservice/update_route_calculator.html} for full documentation.
}
\keyword{internal}
