% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mq_operations.R
\name{mq_create_tags}
\alias{mq_create_tags}
\title{Add a tag to a resource}
\usage{
mq_create_tags(ResourceArn, Tags = NULL)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource tag.}

\item{Tags}{The key-value pair for the resource tag.}
}
\description{
Add a tag to a resource.

See \url{https://www.paws-r-sdk.com/docs/mq_create_tags/} for full documentation.
}
\keyword{internal}
