% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mq_operations.R
\name{mq_delete_broker}
\alias{mq_delete_broker}
\title{Deletes a broker}
\usage{
mq_delete_broker(BrokerId)
}
\arguments{
\item{BrokerId}{[required] The unique ID that Amazon MQ generates for the broker.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  BrokerId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Deletes a broker. Note: This API is asynchronous.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_broker(
  BrokerId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
