% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mq_operations.R
\name{mq_delete_configuration}
\alias{mq_delete_configuration}
\title{Deletes the specified configuration}
\usage{
mq_delete_configuration(ConfigurationId)
}
\arguments{
\item{ConfigurationId}{[required] The unique ID that Amazon MQ generates for the configuration.}
}
\description{
Deletes the specified configuration.

See \url{https://www.paws-r-sdk.com/docs/mq_delete_configuration/} for full documentation.
}
\keyword{internal}
