% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mq_operations.R
\name{mq_delete_user}
\alias{mq_delete_user}
\title{Deletes an ActiveMQ user}
\usage{
mq_delete_user(BrokerId, Username)
}
\arguments{
\item{BrokerId}{[required] The unique ID that Amazon MQ generates for the broker.}

\item{Username}{[required] The username of the ActiveMQ user. This value can contain only
alphanumeric characters, dashes, periods, underscores, and tildes (- .
_ ~). This value must be 2-100 characters long.}
}
\description{
Deletes an ActiveMQ user.

See \url{https://www.paws-r-sdk.com/docs/mq_delete_user/} for full documentation.
}
\keyword{internal}
