% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mq_operations.R
\name{mq_describe_shared_resources}
\alias{mq_describe_shared_resources}
\title{Returns the resources shared to a broker}
\usage{
mq_describe_shared_resources(BrokerId, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{BrokerId}{[required] The unique ID that Amazon MQ generates for the broker.}

\item{MaxResults}{The maximum number of resources that Amazon MQ can return per page (20 by default). This value must be an integer from 5 to 100.}

\item{NextToken}{The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.}
}
\description{
Returns the resources shared to a broker.

See \url{https://www.paws-r-sdk.com/docs/mq_describe_shared_resources/} for full documentation.
}
\keyword{internal}
