% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mq_operations.R
\name{mq_describe_user}
\alias{mq_describe_user}
\title{Returns information about an ActiveMQ user}
\usage{
mq_describe_user(BrokerId, Username)
}
\arguments{
\item{BrokerId}{[required] The unique ID that Amazon MQ generates for the broker.}

\item{Username}{[required] The username of the ActiveMQ user. This value can contain only
alphanumeric characters, dashes, periods, underscores, and tildes (- .
_ ~). This value must be 2-100 characters long.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  BrokerId = "string",
  ConsoleAccess = TRUE|FALSE,
  Groups = list(
    "string"
  ),
  Pending = list(
    ConsoleAccess = TRUE|FALSE,
    Groups = list(
      "string"
    ),
    PendingChange = "CREATE"|"UPDATE"|"DELETE"
  ),
  Username = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns information about an ActiveMQ user.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_user(
  BrokerId = "string",
  Username = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
