% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mq_operations.R
\name{mq_list_tags}
\alias{mq_list_tags}
\title{Lists tags for a resource}
\usage{
mq_list_tags(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource tag.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Lists tags for a resource.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_tags(
  ResourceArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
