% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mq_operations.R
\name{mq_list_users}
\alias{mq_list_users}
\title{Returns a list of all ActiveMQ users}
\usage{
mq_list_users(BrokerId, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{BrokerId}{[required] The unique ID that Amazon MQ generates for the broker.}

\item{MaxResults}{The maximum number of brokers that Amazon MQ can return per page (20 by
default). This value must be an integer from 5 to 100.}

\item{NextToken}{The token that specifies the next page of results Amazon MQ should
return. To request the first page, leave nextToken empty.}
}
\description{
Returns a list of all ActiveMQ users.

See \url{https://www.paws-r-sdk.com/docs/mq_list_users/} for full documentation.
}
\keyword{internal}
