% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mq_operations.R
\name{mq_promote}
\alias{mq_promote}
\title{Promotes a data replication replica broker to the primary broker role}
\usage{
mq_promote(BrokerId, Mode)
}
\arguments{
\item{BrokerId}{[required] The unique ID that Amazon MQ generates for the broker.}

\item{Mode}{[required] The Promote mode requested. Note: Valid values for the parameter are
SWITCHOVER, FAILOVER.}
}
\description{
Promotes a data replication replica broker to the primary broker role.

See \url{https://www.paws-r-sdk.com/docs/mq_promote/} for full documentation.
}
\keyword{internal}
