% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mq_operations.R
\name{mq_reboot_broker}
\alias{mq_reboot_broker}
\title{Reboots a broker}
\usage{
mq_reboot_broker(BrokerId)
}
\arguments{
\item{BrokerId}{[required] The unique ID that Amazon MQ generates for the broker.}
}
\description{
Reboots a broker. Note: This API is asynchronous.

See \url{https://paws-r.github.io/docs/mq/reboot_broker.html} for full documentation.
}
\keyword{internal}
