% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mq_operations.R
\name{mq_update_configuration}
\alias{mq_update_configuration}
\title{Updates the specified configuration}
\usage{
mq_update_configuration(ConfigurationId, Data, Description)
}
\arguments{
\item{ConfigurationId}{[required] The unique ID that Amazon MQ generates for the configuration.}

\item{Data}{Required. The base64-encoded XML configuration.}

\item{Description}{The description of the configuration.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Arn = "string",
  Created = as.POSIXct(
    "2015-01-01"
  ),
  Id = "string",
  LatestRevision = list(
    Created = as.POSIXct(
      "2015-01-01"
    ),
    Description = "string",
    Revision = 123
  ),
  Name = "string",
  Warnings = list(
    list(
      AttributeName = "string",
      ElementName = "string",
      Reason = "DISALLOWED_ELEMENT_REMOVED"|"DISALLOWED_ATTRIBUTE_REMOVED"|"INVALID_ATTRIBUTE_VALUE_REMOVED"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Updates the specified configuration.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_configuration(
  ConfigurationId = "string",
  Data = "string",
  Description = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
