% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwaa_operations.R
\name{mwaa_create_cli_token}
\alias{mwaa_create_cli_token}
\title{Creates a CLI token for the Airflow CLI}
\usage{
mwaa_create_cli_token(Name)
}
\arguments{
\item{Name}{[required] The name of the Amazon MWAA environment. For example,
\code{MyMWAAEnvironment}.}
}
\description{
Creates a CLI token for the Airflow CLI. To learn more, see \href{https://docs.aws.amazon.com/mwaa/latest/userguide/call-mwaa-apis-cli.html}{Creating an Apache Airflow CLI token}.

See \url{https://www.paws-r-sdk.com/docs/mwaa_create_cli_token/} for full documentation.
}
\keyword{internal}
