% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwaa_operations.R
\name{mwaa_create_web_login_token}
\alias{mwaa_create_web_login_token}
\title{Creates a web login token for the Airflow Web UI}
\usage{
mwaa_create_web_login_token(Name)
}
\arguments{
\item{Name}{[required] The name of the Amazon MWAA environment. For example,
\code{MyMWAAEnvironment}.}
}
\description{
Creates a web login token for the Airflow Web UI. To learn more, see \href{https://docs.aws.amazon.com/mwaa/latest/userguide/call-mwaa-apis-web.html}{Creating an Apache Airflow web login token}.

See \url{https://www.paws-r-sdk.com/docs/mwaa_create_web_login_token/} for full documentation.
}
\keyword{internal}
