% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwaa_operations.R
\name{mwaa_list_environments}
\alias{mwaa_list_environments}
\title{Lists the Amazon Managed Workflows for Apache Airflow (MWAA)
environments}
\usage{
mwaa_list_environments(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of results to retrieve per page. For example, \code{5}
environments per page.}

\item{NextToken}{Retrieves the next page of the results.}
}
\description{
Lists the Amazon Managed Workflows for Apache Airflow (MWAA) environments.

See \url{https://www.paws-r-sdk.com/docs/mwaa_list_environments/} for full documentation.
}
\keyword{internal}
