% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwaa_operations.R
\name{mwaa_list_tags_for_resource}
\alias{mwaa_list_tags_for_resource}
\title{Lists the key-value tag pairs associated to the Amazon Managed Workflows
for Apache Airflow (MWAA) environment}
\usage{
mwaa_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the Amazon MWAA environment. For
example,
\code{arn:aws:airflow:us-east-1:123456789012:environment/MyMWAAEnvironment}.}
}
\description{
Lists the key-value tag pairs associated to the Amazon Managed Workflows for Apache Airflow (MWAA) environment. For example, \code{"Environment": "Staging"}.

See \url{https://www.paws-r-sdk.com/docs/mwaa_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
