% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwaa_operations.R
\name{mwaa_publish_metrics}
\alias{mwaa_publish_metrics}
\title{Internal only}
\usage{
mwaa_publish_metrics(EnvironmentName, MetricData)
}
\arguments{
\item{EnvironmentName}{[required] \strong{Internal only}. The name of the environment.}

\item{MetricData}{[required] \strong{Internal only}. Publishes metrics to Amazon CloudWatch. To learn more
about the metrics published to Amazon CloudWatch, see \href{https://docs.aws.amazon.com/mwaa/latest/userguide/cw-metrics.html}{Amazon MWAA performance metrics in Amazon CloudWatch}.}
}
\description{
\strong{Internal only}. Publishes environment health metrics to Amazon CloudWatch.

See \url{https://www.paws-r-sdk.com/docs/mwaa_publish_metrics/} for full documentation.
}
\keyword{internal}
