% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwaa_operations.R
\name{mwaa_tag_resource}
\alias{mwaa_tag_resource}
\title{Associates key-value tag pairs to your Amazon Managed Workflows for
Apache Airflow (MWAA) environment}
\usage{
mwaa_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the Amazon MWAA environment. For
example,
\code{arn:aws:airflow:us-east-1:123456789012:environment/MyMWAAEnvironment}.}

\item{Tags}{[required] The key-value tag pairs you want to associate to your environment. For
example, \code{"Environment": "Staging"}. To learn more, see \href{https://docs.aws.amazon.com/general/latest/gr/aws_tagging.html}{Tagging Amazon Web Services resources}.}
}
\description{
Associates key-value tag pairs to your Amazon Managed Workflows for Apache Airflow (MWAA) environment.

See \url{https://paws-r.github.io/docs/mwaa/tag_resource.html} for full documentation.
}
\keyword{internal}
