% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwaa_operations.R
\name{mwaa_untag_resource}
\alias{mwaa_untag_resource}
\title{Removes key-value tag pairs associated to your Amazon Managed Workflows
for Apache Airflow (MWAA) environment}
\usage{
mwaa_untag_resource(ResourceArn, tagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the Amazon MWAA environment. For
example,
\code{arn:aws:airflow:us-east-1:123456789012:environment/MyMWAAEnvironment}.}

\item{tagKeys}{[required] The key-value tag pair you want to remove. For example,
\code{"Environment": "Staging"}.}
}
\description{
Removes key-value tag pairs associated to your Amazon Managed Workflows for Apache Airflow (MWAA) environment. For example, \code{"Environment": "Staging"}.

See \url{https://paws-r.github.io/docs/mwaa/untag_resource.html} for full documentation.
}
\keyword{internal}
