% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceexplorer_operations.R
\name{resourceexplorer_create_resource_explorer_setup}
\alias{resourceexplorer_create_resource_explorer_setup}
\title{Creates a Resource Explorer setup configuration across multiple Amazon
Web Services Regions}
\usage{
resourceexplorer_create_resource_explorer_setup(
  RegionList,
  AggregatorRegions = NULL,
  ViewName
)
}
\arguments{
\item{RegionList}{[required] A list of Amazon Web Services Regions where Resource Explorer should be
configured. Each Region in the list will have a user-owned index
created.}

\item{AggregatorRegions}{A list of Amazon Web Services Regions that should be configured as
aggregator Regions. Aggregator Regions receive replicated index
information from all other Regions where there is a user-owned index.}

\item{ViewName}{[required] The name for the view to be created as part of the Resource Explorer
setup. The view name must be unique within the Amazon Web Services
account and Region.}
}
\description{
Creates a Resource Explorer setup configuration across multiple Amazon Web Services Regions. This operation sets up indexes and views in the specified Regions. This operation can also be used to set an aggregator Region for cross-Region resource search.

See \url{https://www.paws-r-sdk.com/docs/resourceexplorer_create_resource_explorer_setup/} for full documentation.
}
\keyword{internal}
