% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceexplorer_operations.R
\name{resourceexplorer_delete_resource_explorer_setup}
\alias{resourceexplorer_delete_resource_explorer_setup}
\title{Deletes a Resource Explorer setup configuration}
\usage{
resourceexplorer_delete_resource_explorer_setup(
  RegionList = NULL,
  DeleteInAllRegions = NULL
)
}
\arguments{
\item{RegionList}{A list of Amazon Web Services Regions from which to delete the Resource
Explorer configuration. If not specified, the operation uses the
\code{DeleteInAllRegions} parameter to determine scope.}

\item{DeleteInAllRegions}{Specifies whether to delete Resource Explorer configuration from all
Regions where it is currently enabled. If this parameter is set to
\code{true}, a value for \code{RegionList} must not be provided. Otherwise, the
operation fails with a \code{ValidationException} error.}
}
\description{
Deletes a Resource Explorer setup configuration. This operation removes indexes and views from the specified Regions or all Regions where Resource Explorer is configured.

See \url{https://www.paws-r-sdk.com/docs/resourceexplorer_delete_resource_explorer_setup/} for full documentation.
}
\keyword{internal}
