% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceexplorer_operations.R
\name{resourceexplorer_get_default_view}
\alias{resourceexplorer_get_default_view}
\title{Retrieves the Amazon Resource Name (ARN) of the view that is the default
for the Amazon Web Services Region in which you call this operation}
\usage{
resourceexplorer_get_default_view()
}
\description{
Retrieves the Amazon Resource Name (ARN) of the view that is the default for the Amazon Web Services Region in which you call this operation. You can then call \code{\link[=resourceexplorer_get_view]{get_view}} to retrieve the details of that view.

See \url{https://www.paws-r-sdk.com/docs/resourceexplorer_get_default_view/} for full documentation.
}
\keyword{internal}
