% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceexplorer_operations.R
\name{resourceexplorer_get_index}
\alias{resourceexplorer_get_index}
\title{Retrieves details about the Amazon Web Services Resource Explorer index
in the Amazon Web Services Region in which you invoked the operation}
\usage{
resourceexplorer_get_index()
}
\description{
Retrieves details about the Amazon Web Services Resource Explorer index in the Amazon Web Services Region in which you invoked the operation.

See \url{https://www.paws-r-sdk.com/docs/resourceexplorer_get_index/} for full documentation.
}
\keyword{internal}
