% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceexplorer_operations.R
\name{resourceexplorer_get_managed_view}
\alias{resourceexplorer_get_managed_view}
\title{Retrieves details of the specified Amazon Web Services-managed view}
\usage{
resourceexplorer_get_managed_view(ManagedViewArn)
}
\arguments{
\item{ManagedViewArn}{[required] The Amazon resource name (ARN) of the managed view.}
}
\description{
Retrieves details of the specified \href{https://docs.aws.amazon.com/resource-explorer/latest/userguide/aws-managed-views.html}{Amazon Web Services-managed view}.

See \url{https://www.paws-r-sdk.com/docs/resourceexplorer_get_managed_view/} for full documentation.
}
\keyword{internal}
