% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceexplorer_operations.R
\name{resourceexplorer_get_resource_explorer_setup}
\alias{resourceexplorer_get_resource_explorer_setup}
\title{Retrieves the status and details of a Resource Explorer setup operation}
\usage{
resourceexplorer_get_resource_explorer_setup(
  TaskId,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{TaskId}{[required] The unique identifier of the setup task to retrieve status information
for. This ID is returned by
\code{\link[=resourceexplorer_create_resource_explorer_setup]{create_resource_explorer_setup}}
or
\code{\link[=resourceexplorer_delete_resource_explorer_setup]{delete_resource_explorer_setup}}
operations.}

\item{MaxResults}{The maximum number of Region status results to return in a single
response. Valid values are between \code{1} and \code{100}.}

\item{NextToken}{The pagination token from a previous
\code{\link[=resourceexplorer_get_resource_explorer_setup]{get_resource_explorer_setup}}
response. Use this token to retrieve the next set of results.}
}
\description{
Retrieves the status and details of a Resource Explorer setup operation. This operation returns information about the progress of creating or deleting Resource Explorer configurations across Regions.

See \url{https://www.paws-r-sdk.com/docs/resourceexplorer_get_resource_explorer_setup/} for full documentation.
}
\keyword{internal}
