% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceexplorer_operations.R
\name{resourceexplorer_get_service_index}
\alias{resourceexplorer_get_service_index}
\title{Retrieves information about the Resource Explorer index in the current
Amazon Web Services Region}
\usage{
resourceexplorer_get_service_index()
}
\description{
Retrieves information about the Resource Explorer index in the current Amazon Web Services Region. This operation returns the ARN and type of the index if one exists.

See \url{https://www.paws-r-sdk.com/docs/resourceexplorer_get_service_index/} for full documentation.
}
\keyword{internal}
