% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceexplorer_operations.R
\name{resourceexplorer_get_service_view}
\alias{resourceexplorer_get_service_view}
\title{Retrieves details about a specific Resource Explorer service view}
\usage{
resourceexplorer_get_service_view(ServiceViewArn)
}
\arguments{
\item{ServiceViewArn}{[required] The Amazon Resource Name (ARN) of the service view to retrieve details
for.}
}
\description{
Retrieves details about a specific Resource Explorer service view. This operation returns the configuration and properties of the specified view.

See \url{https://www.paws-r-sdk.com/docs/resourceexplorer_get_service_view/} for full documentation.
}
\keyword{internal}
