% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceexplorer_operations.R
\name{resourceexplorer_list_indexes_for_members}
\alias{resourceexplorer_list_indexes_for_members}
\title{Retrieves a list of a member's indexes in all Amazon Web Services
Regions that are currently collecting resource information for Amazon
Web Services Resource Explorer}
\usage{
resourceexplorer_list_indexes_for_members(
  AccountIdList,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AccountIdList}{[required] The account IDs will limit the output to only indexes from these
accounts.}

\item{MaxResults}{The maximum number of results that you want included on each page of the
response. If you do not include this parameter, it defaults to a value
appropriate to the operation. If additional items exist beyond those
included in the current response, the \code{NextToken} response element is
present and has a value (is not null). Include that value as the
\code{NextToken} request parameter in the next call to the operation to get
the next part of the results.

An API operation can return fewer results than the maximum even when
there are more results available. You should check \code{NextToken} after
every operation to ensure that you receive all of the results.}

\item{NextToken}{The parameter for receiving additional results if you receive a
\code{NextToken} response in a previous request. A \code{NextToken} response
indicates that more output is available. Set this parameter to the value
of the previous call's \code{NextToken} response to indicate where the output
should continue from. The pagination tokens expire after 24 hours.}
}
\description{
Retrieves a list of a member's indexes in all Amazon Web Services Regions that are currently collecting resource information for Amazon Web Services Resource Explorer. Only the management account or a delegated administrator with service access enabled can invoke this API call.

See \url{https://www.paws-r-sdk.com/docs/resourceexplorer_list_indexes_for_members/} for full documentation.
}
\keyword{internal}
