% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceexplorer_operations.R
\name{resourceexplorer_list_service_views}
\alias{resourceexplorer_list_service_views}
\title{Lists all Resource Explorer service views available in the current
Amazon Web Services account}
\usage{
resourceexplorer_list_service_views(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of service view results to return in a single
response. Valid values are between \code{1} and \code{50}.}

\item{NextToken}{The pagination token from a previous
\code{\link[=resourceexplorer_list_service_views]{list_service_views}} response.
Use this token to retrieve the next set of results.}
}
\description{
Lists all Resource Explorer service views available in the current Amazon Web Services account. This operation returns the ARNs of available service views.

See \url{https://www.paws-r-sdk.com/docs/resourceexplorer_list_service_views/} for full documentation.
}
\keyword{internal}
