% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceexplorer_operations.R
\name{resourceexplorer_list_tags_for_resource}
\alias{resourceexplorer_list_tags_for_resource}
\title{Lists the tags that are attached to the specified resource}
\usage{
resourceexplorer_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html?refid=e072b57e-c748-4712-a675-b2b5d0ef7f14&trkcampaign=tw-webinar-series}{Amazon resource name (ARN)} of the view or index that you want to attach tags to.}
}
\description{
Lists the tags that are attached to the specified resource.

See \url{https://www.paws-r-sdk.com/docs/resourceexplorer_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
