% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceexplorer_operations.R
\name{resourceexplorer_tag_resource}
\alias{resourceexplorer_tag_resource}
\title{Adds one or more tag key and value pairs to an Amazon Web Services
Resource Explorer view or index}
\usage{
resourceexplorer_tag_resource(resourceArn, Tags = NULL)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the view or index that you want to
attach tags to.}

\item{Tags}{A list of tag key and value pairs that you want to attach to the
specified view or index.}
}
\description{
Adds one or more tag key and value pairs to an Amazon Web Services Resource Explorer view or index.

See \url{https://www.paws-r-sdk.com/docs/resourceexplorer_tag_resource/} for full documentation.
}
\keyword{internal}
