% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceexplorer_operations.R
\name{resourceexplorer_untag_resource}
\alias{resourceexplorer_untag_resource}
\title{Removes one or more tag key and value pairs from an Amazon Web Services
Resource Explorer view or index}
\usage{
resourceexplorer_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the view or index that you want to
remove tags from.}

\item{tagKeys}{[required] A list of the keys for the tags that you want to remove from the
specified view or index.}
}
\description{
Removes one or more tag key and value pairs from an Amazon Web Services Resource Explorer view or index.

See \url{https://www.paws-r-sdk.com/docs/resourceexplorer_untag_resource/} for full documentation.
}
\keyword{internal}
