% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceexplorer_operations.R
\name{resourceexplorer_update_index_type}
\alias{resourceexplorer_update_index_type}
\title{Changes the type of the index from one of the following types to the
other}
\usage{
resourceexplorer_update_index_type(Arn, Type)
}
\arguments{
\item{Arn}{[required] The \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html?refid=e072b57e-c748-4712-a675-b2b5d0ef7f14&trkcampaign=tw-webinar-series}{Amazon resource name (ARN)} of the index that you want to update.}

\item{Type}{[required] The type of the index. To understand the difference between \code{LOCAL} and \code{AGGREGATOR}, see \href{https://docs.aws.amazon.com/resource-explorer/latest/userguide/manage-aggregator-region.html}{Turning on cross-Region search} in the \emph{Amazon Web Services Resource Explorer User Guide}.}
}
\description{
Changes the type of the index from one of the following types to the other. For more information about indexes and the role they perform in Amazon Web Services Resource Explorer, see \href{https://docs.aws.amazon.com/resource-explorer/latest/userguide/manage-aggregator-region.html}{Turning on cross-Region search by creating an aggregator index} in the \emph{Amazon Web Services Resource Explorer User Guide}.

See \url{https://www.paws-r-sdk.com/docs/resourceexplorer_update_index_type/} for full documentation.
}
\keyword{internal}
