% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_create_discoverer}
\alias{schemas_create_discoverer}
\title{Creates a discoverer}
\usage{
schemas_create_discoverer(
  Description = NULL,
  SourceArn,
  CrossAccount = NULL,
  Tags = NULL
)
}
\arguments{
\item{Description}{A description for the discoverer.}

\item{SourceArn}{[required] The ARN of the event bus.}

\item{CrossAccount}{Support discovery of schemas in events sent to the bus from another
account. (default: true).}

\item{Tags}{Tags associated with the resource.}
}
\description{
Creates a discoverer.

See \url{https://paws-r.github.io/docs/schemas/create_discoverer.html} for full documentation.
}
\keyword{internal}
