% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_create_registry}
\alias{schemas_create_registry}
\title{Creates a registry}
\usage{
schemas_create_registry(Description = NULL, RegistryName, Tags = NULL)
}
\arguments{
\item{Description}{A description of the registry to be created.}

\item{RegistryName}{[required] The name of the registry.}

\item{Tags}{Tags to associate with the registry.}
}
\description{
Creates a registry.

See \url{https://paws-r.github.io/docs/schemas/create_registry.html} for full documentation.
}
\keyword{internal}
