% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_create_schema}
\alias{schemas_create_schema}
\title{Creates a schema definition}
\usage{
schemas_create_schema(
  Content,
  Description = NULL,
  RegistryName,
  SchemaName,
  Tags = NULL,
  Type
)
}
\arguments{
\item{Content}{[required] The source of the schema definition.}

\item{Description}{A description of the schema.}

\item{RegistryName}{[required] The name of the registry.}

\item{SchemaName}{[required] The name of the schema.}

\item{Tags}{Tags associated with the schema.}

\item{Type}{[required] The type of schema.}
}
\description{
Creates a schema definition.

See \url{https://paws-r.github.io/docs/schemas/create_schema.html} for full documentation.
}
\keyword{internal}
