% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_delete_discoverer}
\alias{schemas_delete_discoverer}
\title{Deletes a discoverer}
\usage{
schemas_delete_discoverer(DiscovererId)
}
\arguments{
\item{DiscovererId}{[required] The ID of the discoverer.}
}
\description{
Deletes a discoverer.

See \url{https://paws-r.github.io/docs/schemas/delete_discoverer.html} for full documentation.
}
\keyword{internal}
