% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_delete_resource_policy}
\alias{schemas_delete_resource_policy}
\title{Delete the resource-based policy attached to the specified registry}
\usage{
schemas_delete_resource_policy(RegistryName = NULL)
}
\arguments{
\item{RegistryName}{The name of the registry.}
}
\description{
Delete the resource-based policy attached to the specified registry.

See \url{https://paws-r.github.io/docs/schemas/delete_resource_policy.html} for full documentation.
}
\keyword{internal}
