% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_delete_schema}
\alias{schemas_delete_schema}
\title{Delete a schema definition}
\usage{
schemas_delete_schema(RegistryName, SchemaName)
}
\arguments{
\item{RegistryName}{[required] The name of the registry.}

\item{SchemaName}{[required] The name of the schema.}
}
\description{
Delete a schema definition.

See \url{https://paws-r.github.io/docs/schemas/delete_schema.html} for full documentation.
}
\keyword{internal}
