% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_delete_schema_version}
\alias{schemas_delete_schema_version}
\title{Delete the schema version definition}
\usage{
schemas_delete_schema_version(RegistryName, SchemaName, SchemaVersion)
}
\arguments{
\item{RegistryName}{[required] The name of the registry.}

\item{SchemaName}{[required] The name of the schema.}

\item{SchemaVersion}{[required] The version number of the schema}
}
\description{
Delete the schema version definition

See \url{https://www.paws-r-sdk.com/docs/schemas_delete_schema_version/} for full documentation.
}
\keyword{internal}
