% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_describe_discoverer}
\alias{schemas_describe_discoverer}
\title{Describes the discoverer}
\usage{
schemas_describe_discoverer(DiscovererId)
}
\arguments{
\item{DiscovererId}{[required] The ID of the discoverer.}
}
\description{
Describes the discoverer.

See \url{https://www.paws-r-sdk.com/docs/schemas_describe_discoverer/} for full documentation.
}
\keyword{internal}
