% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_export_schema}
\alias{schemas_export_schema}
\title{Export schema}
\usage{
schemas_export_schema(RegistryName, SchemaName, SchemaVersion = NULL, Type)
}
\arguments{
\item{RegistryName}{[required] The name of the registry.}

\item{SchemaName}{[required] The name of the schema.}

\item{SchemaVersion}{Specifying this limits the results to only this schema version.}

\item{Type}{[required]}
}
\description{
Export schema

See \url{https://www.paws-r-sdk.com/docs/schemas_export_schema/} for full documentation.
}
\keyword{internal}
