% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_get_resource_policy}
\alias{schemas_get_resource_policy}
\title{Retrieves the resource-based policy attached to a given registry}
\usage{
schemas_get_resource_policy(RegistryName = NULL)
}
\arguments{
\item{RegistryName}{The name of the registry.}
}
\description{
Retrieves the resource-based policy attached to a given registry.

See \url{https://paws-r.github.io/docs/schemas/get_resource_policy.html} for full documentation.
}
\keyword{internal}
