% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_list_registries}
\alias{schemas_list_registries}
\title{List the registries}
\usage{
schemas_list_registries(
  Limit = NULL,
  NextToken = NULL,
  RegistryNamePrefix = NULL,
  Scope = NULL
)
}
\arguments{
\item{Limit}{}

\item{NextToken}{The token that specifies the next page of results to return. To request
the first page, leave NextToken empty. The token will expire in 24
hours, and cannot be shared with other accounts.}

\item{RegistryNamePrefix}{Specifying this limits the results to only those registry names that
start with the specified prefix.}

\item{Scope}{Can be set to Local or AWS to limit responses to your custom registries,
or the ones provided by AWS.}
}
\description{
List the registries.

See \url{https://www.paws-r-sdk.com/docs/schemas_list_registries/} for full documentation.
}
\keyword{internal}
