% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_list_schema_versions}
\alias{schemas_list_schema_versions}
\title{Provides a list of the schema versions and related information}
\usage{
schemas_list_schema_versions(
  Limit = NULL,
  NextToken = NULL,
  RegistryName,
  SchemaName
)
}
\arguments{
\item{Limit}{}

\item{NextToken}{The token that specifies the next page of results to return. To request
the first page, leave NextToken empty. The token will expire in 24
hours, and cannot be shared with other accounts.}

\item{RegistryName}{[required] The name of the registry.}

\item{SchemaName}{[required] The name of the schema.}
}
\description{
Provides a list of the schema versions and related information.

See \url{https://paws-r.github.io/docs/schemas/list_schema_versions.html} for full documentation.
}
\keyword{internal}
