% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_list_tags_for_resource}
\alias{schemas_list_tags_for_resource}
\title{Get tags for resource}
\usage{
schemas_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the resource.}
}
\description{
Get tags for resource.

See \url{https://www.paws-r-sdk.com/docs/schemas_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
