% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_put_resource_policy}
\alias{schemas_put_resource_policy}
\title{The name of the policy}
\usage{
schemas_put_resource_policy(Policy, RegistryName = NULL, RevisionId = NULL)
}
\arguments{
\item{Policy}{[required] The resource-based policy.}

\item{RegistryName}{The name of the registry.}

\item{RevisionId}{The revision ID of the policy.}
}
\description{
The name of the policy.

See \url{https://paws-r.github.io/docs/schemas/put_resource_policy.html} for full documentation.
}
\keyword{internal}
