% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_search_schemas}
\alias{schemas_search_schemas}
\title{Search the schemas}
\usage{
schemas_search_schemas(Keywords, Limit = NULL, NextToken = NULL, RegistryName)
}
\arguments{
\item{Keywords}{[required] Specifying this limits the results to only schemas that include the
provided keywords.}

\item{Limit}{}

\item{NextToken}{The token that specifies the next page of results to return. To request
the first page, leave NextToken empty. The token will expire in 24
hours, and cannot be shared with other accounts.}

\item{RegistryName}{[required] The name of the registry.}
}
\description{
Search the schemas

See \url{https://paws-r.github.io/docs/schemas/search_schemas.html} for full documentation.
}
\keyword{internal}
