% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_start_discoverer}
\alias{schemas_start_discoverer}
\title{Starts the discoverer}
\usage{
schemas_start_discoverer(DiscovererId)
}
\arguments{
\item{DiscovererId}{[required] The ID of the discoverer.}
}
\description{
Starts the discoverer

See \url{https://paws-r.github.io/docs/schemas/start_discoverer.html} for full documentation.
}
\keyword{internal}
