% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_stop_discoverer}
\alias{schemas_stop_discoverer}
\title{Stops the discoverer}
\usage{
schemas_stop_discoverer(DiscovererId)
}
\arguments{
\item{DiscovererId}{[required] The ID of the discoverer.}
}
\description{
Stops the discoverer

See \url{https://www.paws-r-sdk.com/docs/schemas_stop_discoverer/} for full documentation.
}
\keyword{internal}
