% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_tag_resource}
\alias{schemas_tag_resource}
\title{Add tags to a resource}
\usage{
schemas_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the resource.}

\item{Tags}{[required] Tags associated with the resource.}
}
\description{
Add tags to a resource.

See \url{https://paws-r.github.io/docs/schemas/tag_resource.html} for full documentation.
}
\keyword{internal}
