% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_untag_resource}
\alias{schemas_untag_resource}
\title{Removes tags from a resource}
\usage{
schemas_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the resource.}

\item{TagKeys}{[required] Keys of key-value pairs.}
}
\description{
Removes tags from a resource.

See \url{https://www.paws-r-sdk.com/docs/schemas_untag_resource/} for full documentation.
}
\keyword{internal}
