% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_update_discoverer}
\alias{schemas_update_discoverer}
\title{Updates the discoverer}
\usage{
schemas_update_discoverer(
  Description = NULL,
  DiscovererId,
  CrossAccount = NULL
)
}
\arguments{
\item{Description}{The description of the discoverer to update.}

\item{DiscovererId}{[required] The ID of the discoverer.}

\item{CrossAccount}{Support discovery of schemas in events sent to the bus from another
account. (default: true)}
}
\description{
Updates the discoverer

See \url{https://paws-r.github.io/docs/schemas/update_discoverer.html} for full documentation.
}
\keyword{internal}
