% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_update_registry}
\alias{schemas_update_registry}
\title{Updates a registry}
\usage{
schemas_update_registry(Description = NULL, RegistryName)
}
\arguments{
\item{Description}{The description of the registry to update.}

\item{RegistryName}{[required] The name of the registry.}
}
\description{
Updates a registry.

See \url{https://paws-r.github.io/docs/schemas/update_registry.html} for full documentation.
}
\keyword{internal}
