% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_update_schema}
\alias{schemas_update_schema}
\title{Updates the schema definition}
\usage{
schemas_update_schema(
  ClientTokenId = NULL,
  Content = NULL,
  Description = NULL,
  RegistryName,
  SchemaName,
  Type = NULL
)
}
\arguments{
\item{ClientTokenId}{The ID of the client token.}

\item{Content}{The source of the schema definition.}

\item{Description}{The description of the schema.}

\item{RegistryName}{[required] The name of the registry.}

\item{SchemaName}{[required] The name of the schema.}

\item{Type}{The schema type for the events schema.}
}
\description{
Updates the schema definition

See \url{https://www.paws-r-sdk.com/docs/schemas_update_schema/} for full documentation.
}
\keyword{internal}
