% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_operations.R
\name{sfn_delete_state_machine_alias}
\alias{sfn_delete_state_machine_alias}
\title{Deletes a state machine alias}
\usage{
sfn_delete_state_machine_alias(stateMachineAliasArn)
}
\arguments{
\item{stateMachineAliasArn}{[required] The Amazon Resource Name (ARN) of the state machine alias to delete.}
}
\description{
Deletes a state machine \href{https://docs.aws.amazon.com/step-functions/latest/dg/concepts-state-machine-alias.html}{alias}.

See \url{https://www.paws-r-sdk.com/docs/sfn_delete_state_machine_alias/} for full documentation.
}
\keyword{internal}
