% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_operations.R
\name{sfn_delete_state_machine_version}
\alias{sfn_delete_state_machine_version}
\title{Deletes a state machine version}
\usage{
sfn_delete_state_machine_version(stateMachineVersionArn)
}
\arguments{
\item{stateMachineVersionArn}{[required] The Amazon Resource Name (ARN) of the state machine version to delete.}
}
\description{
Deletes a state machine \href{https://docs.aws.amazon.com/step-functions/latest/dg/concepts-state-machine-version.html}{version}. After you delete a version, you can't call \code{\link[=sfn_start_execution]{start_execution}} using that version's ARN or use the version with a state machine \href{https://docs.aws.amazon.com/step-functions/latest/dg/concepts-state-machine-alias.html}{alias}.

See \url{https://www.paws-r-sdk.com/docs/sfn_delete_state_machine_version/} for full documentation.
}
\keyword{internal}
