% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_operations.R
\name{sfn_describe_map_run}
\alias{sfn_describe_map_run}
\title{Provides information about a Map Run's configuration, progress, and
results}
\usage{
sfn_describe_map_run(mapRunArn)
}
\arguments{
\item{mapRunArn}{[required] The Amazon Resource Name (ARN) that identifies a Map Run.}
}
\description{
Provides information about a Map Run's configuration, progress, and results. For more information, see \href{https://docs.aws.amazon.com/step-functions/latest/dg/concepts-examine-map-run.html}{Examining Map Run} in the \emph{Step Functions Developer Guide}.

See \url{https://paws-r.github.io/docs/sfn/describe_map_run.html} for full documentation.
}
\keyword{internal}
