% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_operations.R
\name{sfn_describe_state_machine_for_execution}
\alias{sfn_describe_state_machine_for_execution}
\title{Describes the state machine associated with a specific execution}
\usage{
sfn_describe_state_machine_for_execution(executionArn)
}
\arguments{
\item{executionArn}{[required] The Amazon Resource Name (ARN) of the execution you want state machine
information for.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  stateMachineArn = "string",
  name = "string",
  definition = "string",
  roleArn = "string",
  updateDate = as.POSIXct(
    "2015-01-01"
  ),
  loggingConfiguration = list(
    level = "ALL"|"ERROR"|"FATAL"|"OFF",
    includeExecutionData = TRUE|FALSE,
    destinations = list(
      list(
        cloudWatchLogsLogGroup = list(
          logGroupArn = "string"
        )
      )
    )
  ),
  tracingConfiguration = list(
    enabled = TRUE|FALSE
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes the state machine associated with a specific execution.

This operation is eventually consistent. The results are best effort and
may not reflect very recent updates and changes.

This API action is not supported by \code{EXPRESS} state machines.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_state_machine_for_execution(
  executionArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
