% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_operations.R
\name{sfn_list_tags_for_resource}
\alias{sfn_list_tags_for_resource}
\title{List tags for a given resource}
\usage{
sfn_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) for the Step Functions state machine or
activity.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
List tags for a given resource.

Tags may only contain Unicode letters, digits, white space, or these
symbols: \verb{_ . : / = + - @}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_tags_for_resource(
  resourceArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
