% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_operations.R
\name{sfn_untag_resource}
\alias{sfn_untag_resource}
\title{Remove a tag from a Step Functions resource}
\usage{
sfn_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) for the Step Functions state machine or
activity.}

\item{tagKeys}{[required] The list of tags to remove from the resource.}
}
\value{
An empty list.
}
\description{
Remove a tag from a Step Functions resource
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
