% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_operations.R
\name{sfn_update_map_run}
\alias{sfn_update_map_run}
\title{Updates an in-progress Map Run's configuration to include changes to the
settings that control maximum concurrency and Map Run failure}
\usage{
sfn_update_map_run(
  mapRunArn,
  maxConcurrency = NULL,
  toleratedFailurePercentage = NULL,
  toleratedFailureCount = NULL
)
}
\arguments{
\item{mapRunArn}{[required] The Amazon Resource Name (ARN) of a Map Run.}

\item{maxConcurrency}{The maximum number of child workflow executions that can be specified to
run in parallel for the Map Run at the same time.}

\item{toleratedFailurePercentage}{The maximum percentage of failed items before the Map Run fails.}

\item{toleratedFailureCount}{The maximum number of failed items before the Map Run fails.}
}
\description{
Updates an in-progress Map Run's configuration to include changes to the settings that control maximum concurrency and Map Run failure.

See \url{https://www.paws-r-sdk.com/docs/sfn_update_map_run/} for full documentation.
}
\keyword{internal}
