% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_operations.R
\name{sns_create_sms_sandbox_phone_number}
\alias{sns_create_sms_sandbox_phone_number}
\title{Adds a destination phone number to an Amazon Web Services account in the
SMS sandbox and sends a one-time password (OTP) to that phone number}
\usage{
sns_create_sms_sandbox_phone_number(PhoneNumber, LanguageCode = NULL)
}
\arguments{
\item{PhoneNumber}{[required] The destination phone number to verify. On verification, Amazon SNS adds
this phone number to the list of verified phone numbers that you can
send SMS messages to.}

\item{LanguageCode}{The language to use for sending the OTP. The default value is \code{en-US}.}
}
\description{
Adds a destination phone number to an Amazon Web Services account in the SMS sandbox and sends a one-time password (OTP) to that phone number.

See \url{https://www.paws-r-sdk.com/docs/sns_create_sms_sandbox_phone_number/} for full documentation.
}
\keyword{internal}
