% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_operations.R
\name{sns_delete_sms_sandbox_phone_number}
\alias{sns_delete_sms_sandbox_phone_number}
\title{Deletes an Amazon Web Services account's verified or pending phone
number from the SMS sandbox}
\usage{
sns_delete_sms_sandbox_phone_number(PhoneNumber)
}
\arguments{
\item{PhoneNumber}{[required] The destination phone number to delete.}
}
\description{
Deletes an Amazon Web Services account's verified or pending phone number from the SMS sandbox.

See \url{https://paws-r.github.io/docs/sns/delete_sms_sandbox_phone_number.html} for full documentation.
}
\keyword{internal}
