% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_operations.R
\name{sns_get_endpoint_attributes}
\alias{sns_get_endpoint_attributes}
\title{Retrieves the endpoint attributes for a device on one of the supported
push notification services, such as GCM (Firebase Cloud Messaging) and
APNS}
\usage{
sns_get_endpoint_attributes(EndpointArn)
}
\arguments{
\item{EndpointArn}{[required] EndpointArn for GetEndpointAttributes input.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Attributes = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the endpoint attributes for a device on one of the supported
push notification services, such as GCM (Firebase Cloud Messaging) and
APNS. For more information, see \href{https://docs.aws.amazon.com/sns/latest/dg/sns-mobile-application-as-subscriber.html}{Using Amazon SNS Mobile Push Notifications}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_endpoint_attributes(
  EndpointArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
