% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_operations.R
\name{sns_get_platform_application_attributes}
\alias{sns_get_platform_application_attributes}
\title{Retrieves the attributes of the platform application object for the
supported push notification services, such as APNS and GCM (Firebase
Cloud Messaging)}
\usage{
sns_get_platform_application_attributes(PlatformApplicationArn)
}
\arguments{
\item{PlatformApplicationArn}{[required] PlatformApplicationArn for GetPlatformApplicationAttributesInput.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Attributes = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the attributes of the platform application object for the
supported push notification services, such as APNS and GCM (Firebase
Cloud Messaging). For more information, see \href{https://docs.aws.amazon.com/sns/latest/dg/sns-mobile-application-as-subscriber.html}{Using Amazon SNS Mobile Push Notifications}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_platform_application_attributes(
  PlatformApplicationArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
