% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_operations.R
\name{sns_get_subscription_attributes}
\alias{sns_get_subscription_attributes}
\title{Returns all of the properties of a subscription}
\usage{
sns_get_subscription_attributes(SubscriptionArn)
}
\arguments{
\item{SubscriptionArn}{[required] The ARN of the subscription whose properties you want to get.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Attributes = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns all of the properties of a subscription.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_subscription_attributes(
  SubscriptionArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
