% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_operations.R
\name{sns_list_origination_numbers}
\alias{sns_list_origination_numbers}
\title{Lists the calling Amazon Web Services account's dedicated origination
numbers and their metadata}
\usage{
sns_list_origination_numbers(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{Token that the previous
\code{\link[=sns_list_origination_numbers]{list_origination_numbers}} request
returns.}

\item{MaxResults}{The maximum number of origination numbers to return.}
}
\description{
Lists the calling Amazon Web Services account's dedicated origination numbers and their metadata. For more information about origination numbers, see \href{https://docs.aws.amazon.com/sns/latest/dg/channels-sms-originating-identities-origination-numbers.html}{Origination numbers} in the \emph{Amazon SNS Developer Guide}.

See \url{https://paws-r.github.io/docs/sns/list_origination_numbers.html} for full documentation.
}
\keyword{internal}
