% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_operations.R
\name{sns_list_phone_numbers_opted_out}
\alias{sns_list_phone_numbers_opted_out}
\title{Returns a list of phone numbers that are opted out, meaning you cannot
send SMS messages to them}
\usage{
sns_list_phone_numbers_opted_out(nextToken = NULL)
}
\arguments{
\item{nextToken}{A \code{NextToken} string is used when you call the
\code{\link[=sns_list_phone_numbers_opted_out]{list_phone_numbers_opted_out}}
action to retrieve additional records that are available after the first
page of results.}
}
\description{
Returns a list of phone numbers that are opted out, meaning you cannot send SMS messages to them.

See \url{https://paws-r.github.io/docs/sns/list_phone_numbers_opted_out.html} for full documentation.
}
\keyword{internal}
