% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_operations.R
\name{sns_list_platform_applications}
\alias{sns_list_platform_applications}
\title{Lists the platform application objects for the supported push
notification services, such as APNS and GCM (Firebase Cloud Messaging)}
\usage{
sns_list_platform_applications(NextToken = NULL)
}
\arguments{
\item{NextToken}{NextToken string is used when calling ListPlatformApplications action to
retrieve additional records that are available after the first page
results.}
}
\description{
Lists the platform application objects for the supported push notification services, such as APNS and GCM (Firebase Cloud Messaging). The results for \code{\link[=sns_list_platform_applications]{list_platform_applications}} are paginated and return a limited list of applications, up to 100. If additional records are available after the first page results, then a NextToken string will be returned. To receive the next page, you call \code{\link[=sns_list_platform_applications]{list_platform_applications}} using the NextToken string received from the previous call. When there are no more records to return, \code{NextToken} will be null. For more information, see \href{https://docs.aws.amazon.com/sns/latest/dg/sns-mobile-application-as-subscriber.html}{Using Amazon SNS Mobile Push Notifications}.

See \url{https://paws-r.github.io/docs/sns/list_platform_applications.html} for full documentation.
}
\keyword{internal}
