% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_operations.R
\name{sns_list_sms_sandbox_phone_numbers}
\alias{sns_list_sms_sandbox_phone_numbers}
\title{Lists the calling Amazon Web Services account's current verified and
pending destination phone numbers in the SMS sandbox}
\usage{
sns_list_sms_sandbox_phone_numbers(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{Token that the previous \code{ListSMSSandboxPhoneNumbersInput} request
returns.}

\item{MaxResults}{The maximum number of phone numbers to return.}
}
\description{
Lists the calling Amazon Web Services account's current verified and pending destination phone numbers in the SMS sandbox.

See \url{https://www.paws-r-sdk.com/docs/sns_list_sms_sandbox_phone_numbers/} for full documentation.
}
\keyword{internal}
