% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_operations.R
\name{sns_list_topics}
\alias{sns_list_topics}
\title{Returns a list of the requester's topics}
\usage{
sns_list_topics(NextToken = NULL)
}
\arguments{
\item{NextToken}{Token returned by the previous \code{\link[=sns_list_topics]{list_topics}} request.}
}
\description{
Returns a list of the requester's topics. Each call returns a limited list of topics, up to 100. If there are more topics, a \code{NextToken} is also returned. Use the \code{NextToken} parameter in a new \code{\link[=sns_list_topics]{list_topics}} call to get further results.

See \url{https://www.paws-r-sdk.com/docs/sns_list_topics/} for full documentation.
}
\keyword{internal}
