% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_operations.R
\name{sns_publish_batch}
\alias{sns_publish_batch}
\title{Publishes up to ten messages to the specified topic}
\usage{
sns_publish_batch(TopicArn, PublishBatchRequestEntries)
}
\arguments{
\item{TopicArn}{[required] The Amazon resource name (ARN) of the topic you want to batch publish
to.}

\item{PublishBatchRequestEntries}{[required] A list of \code{\link[=sns_publish_batch]{publish_batch}} request entries to be
sent to the SNS topic.}
}
\description{
Publishes up to ten messages to the specified topic. This is a batch version of \code{\link[=sns_publish]{publish}}. For FIFO topics, multiple messages within a single batch are published in the order they are sent, and messages are deduplicated within the batch and across batches for 5 minutes.

See \url{https://paws-r.github.io/docs/sns/publish_batch.html} for full documentation.
}
\keyword{internal}
