% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqs_operations.R
\name{sqs_cancel_message_move_task}
\alias{sqs_cancel_message_move_task}
\title{Cancels a specified message movement task}
\usage{
sqs_cancel_message_move_task(TaskHandle)
}
\arguments{
\item{TaskHandle}{[required] An identifier associated with a message movement task.}
}
\description{
Cancels a specified message movement task. A message movement can only be cancelled when the current status is RUNNING. Cancelling a message movement task does not revert the messages that have already been moved. It can only stop the messages that have not been moved yet.

See \url{https://www.paws-r-sdk.com/docs/sqs_cancel_message_move_task/} for full documentation.
}
\keyword{internal}
