% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqs_operations.R
\name{sqs_delete_message_batch}
\alias{sqs_delete_message_batch}
\title{Deletes up to ten messages from the specified queue}
\usage{
sqs_delete_message_batch(QueueUrl, Entries)
}
\arguments{
\item{QueueUrl}{[required] The URL of the Amazon SQS queue from which messages are deleted.

Queue URLs and names are case-sensitive.}

\item{Entries}{[required] Lists the receipt handles for the messages to be deleted.}
}
\description{
Deletes up to ten messages from the specified queue. This is a batch version of \code{ DeleteMessage.} The result of the action on each message is reported individually in the response.

See \url{https://www.paws-r-sdk.com/docs/sqs_delete_message_batch/} for full documentation.
}
\keyword{internal}
