% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqs_operations.R
\name{sqs_delete_queue}
\alias{sqs_delete_queue}
\title{Deletes the queue specified by the QueueUrl, regardless of the queue's
contents}
\usage{
sqs_delete_queue(QueueUrl)
}
\arguments{
\item{QueueUrl}{[required] The URL of the Amazon SQS queue to delete.

Queue URLs and names are case-sensitive.}
}
\description{
Deletes the queue specified by the \code{QueueUrl}, regardless of the queue's contents.

See \url{https://paws-r.github.io/docs/sqs/delete_queue.html} for full documentation.
}
\keyword{internal}
