% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqs_operations.R
\name{sqs_send_message_batch}
\alias{sqs_send_message_batch}
\title{Delivers up to ten messages to the specified queue}
\usage{
sqs_send_message_batch(QueueUrl, Entries)
}
\arguments{
\item{QueueUrl}{[required] The URL of the Amazon SQS queue to which batched messages are sent.

Queue URLs and names are case-sensitive.}

\item{Entries}{[required] A list of \code{SendMessageBatchRequestEntry} items.}
}
\description{
Delivers up to ten messages to the specified queue. This is a batch version of \code{ SendMessage.} For a FIFO queue, multiple messages within a single batch are enqueued in the order they are sent.

See \url{https://paws-r.github.io/docs/sqs/send_message_batch.html} for full documentation.
}
\keyword{internal}
