% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqs_operations.R
\name{sqs_untag_queue}
\alias{sqs_untag_queue}
\title{Remove cost allocation tags from the specified Amazon SQS queue}
\usage{
sqs_untag_queue(QueueUrl, TagKeys)
}
\arguments{
\item{QueueUrl}{[required] The URL of the queue.}

\item{TagKeys}{[required] The list of tags to be removed from the specified queue.}
}
\description{
Remove cost allocation tags from the specified Amazon SQS queue. For an overview, see \href{https://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/sqs-queue-tags.html}{Tagging Your Amazon SQS Queues} in the \emph{Amazon SQS Developer Guide}.

See \url{https://paws-r.github.io/docs/sqs/untag_queue.html} for full documentation.
}
\keyword{internal}
