% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf_operations.R
\name{swf_count_pending_decision_tasks}
\alias{swf_count_pending_decision_tasks}
\title{Returns the estimated number of decision tasks in the specified task
list}
\usage{
swf_count_pending_decision_tasks(domain, taskList)
}
\arguments{
\item{domain}{[required] The name of the domain that contains the task list.}

\item{taskList}{[required] The name of the task list.}
}
\description{
Returns the estimated number of decision tasks in the specified task list. The count returned is an approximation and isn't guaranteed to be exact. If you specify a task list that no decision task was ever scheduled in then \code{0} is returned.

See \url{https://www.paws-r-sdk.com/docs/swf_count_pending_decision_tasks/} for full documentation.
}
\keyword{internal}
