% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf_operations.R
\name{swf_delete_activity_type}
\alias{swf_delete_activity_type}
\title{Deletes the specified activity type}
\usage{
swf_delete_activity_type(domain, activityType)
}
\arguments{
\item{domain}{[required] The name of the domain in which the activity type is registered.}

\item{activityType}{[required] The activity type to delete.}
}
\description{
Deletes the specified \emph{activity type}.

See \url{https://www.paws-r-sdk.com/docs/swf_delete_activity_type/} for full documentation.
}
\keyword{internal}
