% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf_operations.R
\name{swf_delete_workflow_type}
\alias{swf_delete_workflow_type}
\title{Deletes the specified workflow type}
\usage{
swf_delete_workflow_type(domain, workflowType)
}
\arguments{
\item{domain}{[required] The name of the domain in which the workflow type is registered.}

\item{workflowType}{[required] The workflow type to delete.}
}
\description{
Deletes the specified \emph{workflow type}.

See \url{https://www.paws-r-sdk.com/docs/swf_delete_workflow_type/} for full documentation.
}
\keyword{internal}
