% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf_operations.R
\name{swf_deprecate_domain}
\alias{swf_deprecate_domain}
\title{Deprecates the specified domain}
\usage{
swf_deprecate_domain(name)
}
\arguments{
\item{name}{[required] The name of the domain to deprecate.}
}
\description{
Deprecates the specified domain. After a domain has been deprecated it cannot be used to create new workflow executions or register new types. However, you can still use visibility actions on this domain. Deprecating a domain also deprecates all activity and workflow types registered in the domain. Executions that were started before the domain was deprecated continues to run.

See \url{https://www.paws-r-sdk.com/docs/swf_deprecate_domain/} for full documentation.
}
\keyword{internal}
