% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf_operations.R
\name{swf_describe_activity_type}
\alias{swf_describe_activity_type}
\title{Returns information about the specified activity type}
\usage{
swf_describe_activity_type(domain, activityType)
}
\arguments{
\item{domain}{[required] The name of the domain in which the activity type is registered.}

\item{activityType}{[required] The activity type to get information about. Activity types are
identified by the \code{name} and \code{version} that were supplied when the
activity was registered.}
}
\description{
Returns information about the specified activity type. This includes configuration settings provided when the type was registered and other general information about the type.

See \url{https://www.paws-r-sdk.com/docs/swf_describe_activity_type/} for full documentation.
}
\keyword{internal}
