% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf_operations.R
\name{swf_describe_domain}
\alias{swf_describe_domain}
\title{Returns information about the specified domain, including description
and status}
\usage{
swf_describe_domain(name)
}
\arguments{
\item{name}{[required] The name of the domain to describe.}
}
\description{
Returns information about the specified domain, including description and status.

See \url{https://www.paws-r-sdk.com/docs/swf_describe_domain/} for full documentation.
}
\keyword{internal}
