% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf_operations.R
\name{swf_describe_workflow_execution}
\alias{swf_describe_workflow_execution}
\title{Returns information about the specified workflow execution including its
type and some statistics}
\usage{
swf_describe_workflow_execution(domain, execution)
}
\arguments{
\item{domain}{[required] The name of the domain containing the workflow execution.}

\item{execution}{[required] The workflow execution to describe.}
}
\description{
Returns information about the specified workflow execution including its type and some statistics.

See \url{https://www.paws-r-sdk.com/docs/swf_describe_workflow_execution/} for full documentation.
}
\keyword{internal}
