% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf_operations.R
\name{swf_request_cancel_workflow_execution}
\alias{swf_request_cancel_workflow_execution}
\title{Records a WorkflowExecutionCancelRequested event in the currently
running workflow execution identified by the given domain, workflowId,
and runId}
\usage{
swf_request_cancel_workflow_execution(domain, workflowId, runId = NULL)
}
\arguments{
\item{domain}{[required] The name of the domain containing the workflow execution to cancel.}

\item{workflowId}{[required] The workflowId of the workflow execution to cancel.}

\item{runId}{The runId of the workflow execution to cancel.}
}
\description{
Records a \code{WorkflowExecutionCancelRequested} event in the currently running workflow execution identified by the given domain, workflowId, and runId. This logically requests the cancellation of the workflow execution as a whole. It is up to the decider to take appropriate actions when it receives an execution history with this event.

See \url{https://www.paws-r-sdk.com/docs/swf_request_cancel_workflow_execution/} for full documentation.
}
\keyword{internal}
