% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf_operations.R
\name{swf_respond_activity_task_canceled}
\alias{swf_respond_activity_task_canceled}
\title{Used by workers to tell the service that the ActivityTask identified by
the taskToken was successfully canceled}
\usage{
swf_respond_activity_task_canceled(taskToken, details = NULL)
}
\arguments{
\item{taskToken}{[required] The \code{taskToken} of the ActivityTask.

\code{taskToken} is generated by the service and should be treated as an
opaque value. If the task is passed to another process, its \code{taskToken}
must also be passed. This enables it to provide its progress and respond
with results.}

\item{details}{Information about the cancellation.}
}
\description{
Used by workers to tell the service that the ActivityTask identified by the \code{taskToken} was successfully canceled. Additional \code{details} can be provided using the \code{details} argument.

See \url{https://www.paws-r-sdk.com/docs/swf_respond_activity_task_canceled/} for full documentation.
}
\keyword{internal}
