% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf_operations.R
\name{swf_respond_activity_task_completed}
\alias{swf_respond_activity_task_completed}
\title{Used by workers to tell the service that the ActivityTask identified by
the taskToken completed successfully with a result (if provided)}
\usage{
swf_respond_activity_task_completed(taskToken, result = NULL)
}
\arguments{
\item{taskToken}{[required] The \code{taskToken} of the ActivityTask.

\code{taskToken} is generated by the service and should be treated as an
opaque value. If the task is passed to another process, its \code{taskToken}
must also be passed. This enables it to provide its progress and respond
with results.}

\item{result}{The result of the activity task. It is a free form string that is
implementation specific.}
}
\description{
Used by workers to tell the service that the ActivityTask identified by the \code{taskToken} completed successfully with a \code{result} (if provided). The \code{result} appears in the \code{ActivityTaskCompleted} event in the workflow history.

See \url{https://paws-r.github.io/docs/swf/respond_activity_task_completed.html} for full documentation.
}
\keyword{internal}
