% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf_operations.R
\name{swf_tag_resource}
\alias{swf_tag_resource}
\title{Add a tag to a Amazon SWF domain}
\usage{
swf_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) for the Amazon SWF domain.}

\item{tags}{[required] The list of tags to add to a domain.

Tags may only contain unicode letters, digits, whitespace, or these
symbols: \verb{_ . : / = + - @}.}
}
\value{
An empty list.
}
\description{
Add a tag to a Amazon SWF domain.

Amazon SWF supports a maximum of 50 tags per resource.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  resourceArn = "string",
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}
}

\keyword{internal}
