% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf_operations.R
\name{swf_undeprecate_activity_type}
\alias{swf_undeprecate_activity_type}
\title{Undeprecates a previously deprecated activity type}
\usage{
swf_undeprecate_activity_type(domain, activityType)
}
\arguments{
\item{domain}{[required] The name of the domain of the deprecated activity type.}

\item{activityType}{[required] The activity type to undeprecate.}
}
\description{
Undeprecates a previously deprecated \emph{activity type}. After an activity type has been undeprecated, you can create new tasks of that activity type.

See \url{https://paws-r.github.io/docs/swf/undeprecate_activity_type.html} for full documentation.
}
\keyword{internal}
