% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf_operations.R
\name{swf_undeprecate_domain}
\alias{swf_undeprecate_domain}
\title{Undeprecates a previously deprecated domain}
\usage{
swf_undeprecate_domain(name)
}
\arguments{
\item{name}{[required] The name of the domain of the deprecated workflow type.}
}
\description{
Undeprecates a previously deprecated domain. After a domain has been undeprecated it can be used to create new workflow executions or register new types.

See \url{https://paws-r.github.io/docs/swf/undeprecate_domain.html} for full documentation.
}
\keyword{internal}
