% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf_operations.R
\name{swf_undeprecate_workflow_type}
\alias{swf_undeprecate_workflow_type}
\title{Undeprecates a previously deprecated workflow type}
\usage{
swf_undeprecate_workflow_type(domain, workflowType)
}
\arguments{
\item{domain}{[required] The name of the domain of the deprecated workflow type.}

\item{workflowType}{[required] The name of the domain of the deprecated workflow type.}
}
\description{
Undeprecates a previously deprecated \emph{workflow type}. After a workflow type has been undeprecated, you can create new executions of that type.

See \url{https://www.paws-r-sdk.com/docs/swf_undeprecate_workflow_type/} for full documentation.
}
\keyword{internal}
