% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_cancel_replay}
\alias{eventbridge_cancel_replay}
\title{Cancels the specified replay}
\usage{
eventbridge_cancel_replay(ReplayName)
}
\arguments{
\item{ReplayName}{[required] The name of the replay to cancel.}
}
\description{
Cancels the specified replay.

See \url{https://paws-r.github.io/docs/eventbridge/cancel_replay.html} for full documentation.
}
\keyword{internal}
