% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_deactivate_event_source}
\alias{eventbridge_deactivate_event_source}
\title{An AWS customer uses this operation to temporarily stop receiving events
from the specified partner event source}
\usage{
eventbridge_deactivate_event_source(Name)
}
\arguments{
\item{Name}{[required] The name of the partner event source to deactivate.}
}
\description{
An AWS customer uses this operation to temporarily stop receiving events
from the specified partner event source. The matching event bus isn\'t
deleted.
}
\details{
When you deactivate a partner event source, the source goes into
\code{PENDING} state. If it remains in \code{PENDING} state for more than two
weeks, it\'s deleted.

To activate a deactivated partner event source, use ActivateEventSource.
}
\section{Request syntax}{
\preformatted{svc$deactivate_event_source(
  Name = "string"
)
}
}

\keyword{internal}
