% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_delete_event_bus}
\alias{eventbridge_delete_event_bus}
\title{Deletes the specified custom event bus or partner event bus}
\usage{
eventbridge_delete_event_bus(Name)
}
\arguments{
\item{Name}{[required] The name of the event bus to delete.}
}
\description{
Deletes the specified custom event bus or partner event bus. All rules associated with this event bus need to be deleted. You can't delete your account's default event bus.

See \url{https://paws-r.github.io/docs/eventbridge/delete_event_bus.html} for full documentation.
}
\keyword{internal}
