% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_describe_api_destination}
\alias{eventbridge_describe_api_destination}
\title{Retrieves details about an API destination}
\usage{
eventbridge_describe_api_destination(Name)
}
\arguments{
\item{Name}{[required] The name of the API destination to retrieve.}
}
\description{
Retrieves details about an API destination.

See \url{https://www.paws-r-sdk.com/docs/eventbridge_describe_api_destination/} for full documentation.
}
\keyword{internal}
