% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_describe_archive}
\alias{eventbridge_describe_archive}
\title{Retrieves details about an archive}
\usage{
eventbridge_describe_archive(ArchiveName)
}
\arguments{
\item{ArchiveName}{[required] The name of the archive to retrieve.}
}
\description{
Retrieves details about an archive.
}
\section{Request syntax}{
\preformatted{svc$describe_archive(
  ArchiveName = "string"
)
}
}

\keyword{internal}
