% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_describe_connection}
\alias{eventbridge_describe_connection}
\title{Retrieves details about a connection}
\usage{
eventbridge_describe_connection(Name)
}
\arguments{
\item{Name}{[required] The name of the connection to retrieve.}
}
\description{
Retrieves details about a connection.

See \url{https://paws-r.github.io/docs/eventbridge/describe_connection.html} for full documentation.
}
\keyword{internal}
