% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_describe_event_bus}
\alias{eventbridge_describe_event_bus}
\title{Displays details about an event bus in your account}
\usage{
eventbridge_describe_event_bus(Name = NULL)
}
\arguments{
\item{Name}{The name or ARN of the event bus to show details for. If you omit this,
the default event bus is displayed.}
}
\description{
Displays details about an event bus in your account. This can include the external Amazon Web Services accounts that are permitted to write events to your default event bus, and the associated policy. For custom event buses and partner event buses, it displays the name, ARN, policy, state, and creation time.

See \url{https://www.paws-r-sdk.com/docs/eventbridge_describe_event_bus/} for full documentation.
}
\keyword{internal}
