% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_describe_event_source}
\alias{eventbridge_describe_event_source}
\title{This operation lists details about a partner event source that is shared
with your account}
\usage{
eventbridge_describe_event_source(Name)
}
\arguments{
\item{Name}{[required] The name of the partner event source to display the details of.}
}
\description{
This operation lists details about a partner event source that is shared
with your account.
}
\details{
This operation is run by AWS customers, not by SaaS partners.
}
\section{Request syntax}{
\preformatted{svc$describe_event_source(
  Name = "string"
)
}
}

\keyword{internal}
