% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_describe_partner_event_source}
\alias{eventbridge_describe_partner_event_source}
\title{An SaaS partner can use this operation to list details about a partner
event source that they have created}
\usage{
eventbridge_describe_partner_event_source(Name)
}
\arguments{
\item{Name}{[required] The name of the event source to display.}
}
\description{
An SaaS partner can use this operation to list details about a partner event source that they have created. Amazon Web Services customers do not use this operation. Instead, Amazon Web Services customers can use \code{\link[=eventbridge_describe_event_source]{describe_event_source}} to see details about a partner event source that is shared with them.

See \url{https://paws-r.github.io/docs/eventbridge/describe_partner_event_source.html} for full documentation.
}
\keyword{internal}
