% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_describe_rule}
\alias{eventbridge_describe_rule}
\title{Describes the specified rule}
\usage{
eventbridge_describe_rule(Name, EventBusName)
}
\arguments{
\item{Name}{[required] The name of the rule.}

\item{EventBusName}{The name or ARN of the event bus associated with the rule. If you omit
this, the default event bus is used.}
}
\description{
Describes the specified rule.

DescribeRule does not list the targets of a rule. To see the targets
associated with a rule, use ListTargetsByRule.
}
\section{Request syntax}{
\preformatted{svc$describe_rule(
  Name = "string",
  EventBusName = "string"
)
}
}

\keyword{internal}
