% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_disable_rule}
\alias{eventbridge_disable_rule}
\title{Disables the specified rule}
\usage{
eventbridge_disable_rule(Name, EventBusName = NULL)
}
\arguments{
\item{Name}{[required] The name of the rule.}

\item{EventBusName}{The name or ARN of the event bus associated with the rule. If you omit
this, the default event bus is used.}
}
\description{
Disables the specified rule. A disabled rule won't match any events, and won't self-trigger if it has a schedule expression.

See \url{https://paws-r.github.io/docs/eventbridge/disable_rule.html} for full documentation.
}
\keyword{internal}
