% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_list_api_destinations}
\alias{eventbridge_list_api_destinations}
\title{Retrieves a list of API destination in the account in the current Region}
\usage{
eventbridge_list_api_destinations(
  NamePrefix = NULL,
  ConnectionArn = NULL,
  NextToken = NULL,
  Limit = NULL
)
}
\arguments{
\item{NamePrefix}{A name prefix to filter results returned. Only API destinations with a
name that starts with the prefix are returned.}

\item{ConnectionArn}{The ARN of the connection specified for the API destination.}

\item{NextToken}{The token returned by a previous call to retrieve the next set of
results.}

\item{Limit}{The maximum number of API destinations to include in the response.}
}
\description{
Retrieves a list of API destination in the account in the current Region.

See \url{https://paws-r.github.io/docs/eventbridge/list_api_destinations.html} for full documentation.
}
\keyword{internal}
