% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_list_archives}
\alias{eventbridge_list_archives}
\title{Lists your archives}
\usage{
eventbridge_list_archives(
  NamePrefix = NULL,
  EventSourceArn = NULL,
  State = NULL,
  NextToken = NULL,
  Limit = NULL
)
}
\arguments{
\item{NamePrefix}{A name prefix to filter the archives returned. Only archives with name
that match the prefix are returned.}

\item{EventSourceArn}{The ARN of the event source associated with the archive.}

\item{State}{The state of the archive.}

\item{NextToken}{The token returned by a previous call to retrieve the next set of
results.}

\item{Limit}{The maximum number of results to return.}
}
\description{
Lists your archives. You can either list all the archives or you can provide a prefix to match to the archive names. Filter parameters are exclusive.

See \url{https://paws-r.github.io/docs/eventbridge/list_archives.html} for full documentation.
}
\keyword{internal}
