% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_list_connections}
\alias{eventbridge_list_connections}
\title{Retrieves a list of connections from the account}
\usage{
eventbridge_list_connections(
  NamePrefix = NULL,
  ConnectionState = NULL,
  NextToken = NULL,
  Limit = NULL
)
}
\arguments{
\item{NamePrefix}{A name prefix to filter results returned. Only connections with a name
that starts with the prefix are returned.}

\item{ConnectionState}{The state of the connection.}

\item{NextToken}{The token returned by a previous call to retrieve the next set of
results.}

\item{Limit}{The maximum number of connections to return.}
}
\description{
Retrieves a list of connections from the account.

See \url{https://paws-r.github.io/docs/eventbridge/list_connections.html} for full documentation.
}
\keyword{internal}
