% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_list_partner_event_sources}
\alias{eventbridge_list_partner_event_sources}
\title{An SaaS partner can use this operation to list all the partner event
source names that they have created}
\usage{
eventbridge_list_partner_event_sources(NamePrefix, NextToken, Limit)
}
\arguments{
\item{NamePrefix}{[required] If you specify this, the results are limited to only those partner event
sources that start with the string you specify.}

\item{NextToken}{The token returned by a previous call to this operation. Specifying this
retrieves the next set of results.}

\item{Limit}{pecifying this limits the number of results returned by this operation.
The operation also returns a \code{NextToken} that you can use in a
subsequent operation to retrieve the next set of results.}
}
\description{
An SaaS partner can use this operation to list all the partner event
source names that they have created.
}
\details{
This operation is not used by AWS customers.
}
\section{Request syntax}{
\preformatted{svc$list_partner_event_sources(
  NamePrefix = "string",
  NextToken = "string",
  Limit = 123
)
}
}

\keyword{internal}
