% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_list_rule_names_by_target}
\alias{eventbridge_list_rule_names_by_target}
\title{Lists the rules for the specified target}
\usage{
eventbridge_list_rule_names_by_target(
  TargetArn,
  EventBusName = NULL,
  NextToken = NULL,
  Limit = NULL
)
}
\arguments{
\item{TargetArn}{[required] The Amazon Resource Name (ARN) of the target resource.}

\item{EventBusName}{The name or ARN of the event bus to list rules for. If you omit this,
the default event bus is used.}

\item{NextToken}{The token returned by a previous call to retrieve the next set of
results.}

\item{Limit}{The maximum number of results to return.}
}
\description{
Lists the rules for the specified target. You can see which of the rules in Amazon EventBridge can invoke a specific target in your account.

See \url{https://paws-r.github.io/docs/eventbridge/list_rule_names_by_target.html} for full documentation.
}
\keyword{internal}
