% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_test_event_pattern}
\alias{eventbridge_test_event_pattern}
\title{Tests whether the specified event pattern matches the provided event}
\usage{
eventbridge_test_event_pattern(EventPattern, Event)
}
\arguments{
\item{EventPattern}{[required] The event pattern. For more information, see \href{https://docs.aws.amazon.com/eventbridge/latest/userguide/eb-events.html}{Events and Event Patterns}
in the \emph{Amazon EventBridge User Guide}.}

\item{Event}{[required] The event, in JSON format, to test against the event pattern. The JSON
must follow the format specified in \href{https://docs.aws.amazon.com/eventbridge/latest/userguide/eb-events.html}{Amazon Web Services Events},
and the following fields are mandatory:
\itemize{
\item \code{id}
\item \code{account}
\item \code{source}
\item \code{time}
\item \code{region}
\item \code{resources}
\item \code{detail-type}
}}
}
\description{
Tests whether the specified event pattern matches the provided event.

See \url{https://paws-r.github.io/docs/eventbridge/test_event_pattern.html} for full documentation.
}
\keyword{internal}
