% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_update_archive}
\alias{eventbridge_update_archive}
\title{Updates the specified archive}
\usage{
eventbridge_update_archive(
  ArchiveName,
  Description = NULL,
  EventPattern = NULL,
  RetentionDays = NULL
)
}
\arguments{
\item{ArchiveName}{[required] The name of the archive to update.}

\item{Description}{The description for the archive.}

\item{EventPattern}{The event pattern to use to filter events sent to the archive.}

\item{RetentionDays}{The number of days to retain events in the archive.}
}
\description{
Updates the specified archive.

See \url{https://www.paws-r-sdk.com/docs/eventbridge_update_archive/} for full documentation.
}
\keyword{internal}
