% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbridge_operations.R
\name{eventbridge_update_connection}
\alias{eventbridge_update_connection}
\title{Updates settings for a connection}
\usage{
eventbridge_update_connection(
  Name,
  Description = NULL,
  AuthorizationType = NULL,
  AuthParameters = NULL,
  InvocationConnectivityParameters = NULL
)
}
\arguments{
\item{Name}{[required] The name of the connection to update.}

\item{Description}{A description for the connection.}

\item{AuthorizationType}{The type of authorization to use for the connection.}

\item{AuthParameters}{The authorization parameters to use for the connection.}

\item{InvocationConnectivityParameters}{For connections to private resource endpoints, the parameters to use for
invoking the resource endpoint.

For more information, see \href{https://docs.aws.amazon.com/eventbridge/latest/userguide/}{Connecting to private resources} in
the \emph{Amazon EventBridge User Guide} .}
}
\description{
Updates settings for a connection.

See \url{https://www.paws-r-sdk.com/docs/eventbridge_update_connection/} for full documentation.
}
\keyword{internal}
