% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_batch_delete_device_position_history}
\alias{locationservice_batch_delete_device_position_history}
\title{Deletes the position history of one or more devices from a tracker
resource}
\usage{
locationservice_batch_delete_device_position_history(DeviceIds, TrackerName)
}
\arguments{
\item{DeviceIds}{[required] Devices whose position history you want to delete.
\itemize{
\item For example, for two devices: \verb{“DeviceIds” : [DeviceId1,DeviceId2]}
}}

\item{TrackerName}{[required] The name of the tracker resource to delete the device position history
from.}
}
\description{
Deletes the position history of one or more devices from a tracker resource.

See \url{https://paws-r.github.io/docs/locationservice/batch_delete_device_position_history.html} for full documentation.
}
\keyword{internal}
