% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_delete_geofence_collection}
\alias{locationservice_delete_geofence_collection}
\title{Deletes a geofence collection from your AWS account}
\usage{
locationservice_delete_geofence_collection(CollectionName)
}
\arguments{
\item{CollectionName}{[required] The name of the geofence collection to be deleted.}
}
\description{
Deletes a geofence collection from your AWS account.

See \url{https://paws-r.github.io/docs/locationservice/delete_geofence_collection.html} for full documentation.
}
\keyword{internal}
