% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_delete_key}
\alias{locationservice_delete_key}
\title{Deletes the specified API key}
\usage{
locationservice_delete_key(ForceDelete = NULL, KeyName)
}
\arguments{
\item{ForceDelete}{ForceDelete bypasses an API key's expiry conditions and deletes the key.
Set the parameter \code{true} to delete the key or to \code{false} to not
preemptively delete the API key.

Valid values: \code{true}, or \code{false}.

Required: No

This action is irreversible. Only use ForceDelete if you are certain the
key is no longer in use.}

\item{KeyName}{[required] The name of the API key to delete.}
}
\description{
Deletes the specified API key. The API key must have been deactivated more than 90 days previously.

See \url{https://www.paws-r-sdk.com/docs/locationservice_delete_key/} for full documentation.
}
\keyword{internal}
