% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_delete_route_calculator}
\alias{locationservice_delete_route_calculator}
\title{Deletes a route calculator resource from your Amazon Web Services
account}
\usage{
locationservice_delete_route_calculator(CalculatorName)
}
\arguments{
\item{CalculatorName}{[required] The name of the route calculator resource to be deleted.}
}
\description{
Deletes a route calculator resource from your Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/locationservice_delete_route_calculator/} for full documentation.
}
\keyword{internal}
