% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_delete_tracker}
\alias{locationservice_delete_tracker}
\title{Deletes a tracker resource from your Amazon Web Services account}
\usage{
locationservice_delete_tracker(TrackerName)
}
\arguments{
\item{TrackerName}{[required] The name of the tracker resource to be deleted.}
}
\description{
Deletes a tracker resource from your Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/locationservice_delete_tracker/} for full documentation.
}
\keyword{internal}
