% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_describe_map}
\alias{locationservice_describe_map}
\title{This operation is no longer current and may be deprecated in the future}
\usage{
locationservice_describe_map(MapName)
}
\arguments{
\item{MapName}{[required] The name of the map resource.}
}
\description{
This operation is no longer current and may be deprecated in the future. We recommend upgrading to the Maps API V2 unless you require \code{Grab} data.

See \url{https://www.paws-r-sdk.com/docs/locationservice_describe_map/} for full documentation.
}
\keyword{internal}
