% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_describe_route_calculator}
\alias{locationservice_describe_route_calculator}
\title{Retrieves the route calculator resource details}
\usage{
locationservice_describe_route_calculator(CalculatorName)
}
\arguments{
\item{CalculatorName}{[required] The name of the route calculator resource.}
}
\description{
Retrieves the route calculator resource details.

See \url{https://paws-r.github.io/docs/locationservice/describe_route_calculator.html} for full documentation.
}
\keyword{internal}
