% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_get_map_tile}
\alias{locationservice_get_map_tile}
\title{This operation is no longer current and may be deprecated in the future}
\usage{
locationservice_get_map_tile(MapName, Z, X, Y, Key = NULL)
}
\arguments{
\item{MapName}{[required] The map resource to retrieve the map tiles from.}

\item{Z}{[required] The zoom value for the map tile.}

\item{X}{[required] The X axis value for the map tile.}

\item{Y}{[required] The Y axis value for the map tile.}

\item{Key}{The optional \href{https://docs.aws.amazon.com/location/previous/developerguide/using-apikeys.html}{API key}
to authorize the request.}
}
\description{
This operation is no longer current and may be deprecated in the future. We recommend upgrading to \href{https://docs.aws.amazon.com/location/latest/APIReference/API_geomaps_GetTile.html}{\code{GetTile}} unless you require \code{Grab} data.

See \url{https://www.paws-r-sdk.com/docs/locationservice_get_map_tile/} for full documentation.
}
\keyword{internal}
