% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_list_geofence_collections}
\alias{locationservice_list_geofence_collections}
\title{Lists geofence collections in your Amazon Web Services account}
\usage{
locationservice_list_geofence_collections(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{An optional limit for the number of resources returned in a single call.

Default value: \code{100}}

\item{NextToken}{The pagination token specifying which page of results to return in the
response. If no token is provided, the default page is the first page.

Default value: \code{null}}
}
\description{
Lists geofence collections in your Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/locationservice_list_geofence_collections/} for full documentation.
}
\keyword{internal}
