% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_list_place_indexes}
\alias{locationservice_list_place_indexes}
\title{This operation is no longer current and may be deprecated in the future}
\usage{
locationservice_list_place_indexes(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{An optional limit for the maximum number of results returned in a single
call.

Default value: \code{100}}

\item{NextToken}{The pagination token specifying which page of results to return in the
response. If no token is provided, the default page is the first page.

Default value: \code{null}}
}
\description{
This operation is no longer current and may be deprecated in the future. We recommend you upgrade to the Places API V2 unless you require Grab data.

See \url{https://www.paws-r-sdk.com/docs/locationservice_list_place_indexes/} for full documentation.
}
\keyword{internal}
