% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_list_tags_for_resource}
\alias{locationservice_list_tags_for_resource}
\title{Returns a list of tags that are applied to the specified Amazon Location
resource}
\usage{
locationservice_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource whose tags you want to
retrieve.
\itemize{
\item Format example:
\code{arn:aws:geo:region:account-id:resourcetype/ExampleResource}
}}
}
\description{
Returns a list of tags that are applied to the specified Amazon Location resource.

See \url{https://paws-r.github.io/docs/locationservice/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
