% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_update_geofence_collection}
\alias{locationservice_update_geofence_collection}
\title{Updates the specified properties of a given geofence collection}
\usage{
locationservice_update_geofence_collection(
  CollectionName,
  PricingPlan = NULL,
  PricingPlanDataSource = NULL,
  Description = NULL
)
}
\arguments{
\item{CollectionName}{[required] The name of the geofence collection to update.}

\item{PricingPlan}{No longer used. If included, the only allowed value is
\code{RequestBasedUsage}.}

\item{PricingPlanDataSource}{This parameter is no longer used.}

\item{Description}{Updates the description for the geofence collection.}
}
\description{
Updates the specified properties of a given geofence collection.

See \url{https://www.paws-r-sdk.com/docs/locationservice_update_geofence_collection/} for full documentation.
}
\keyword{internal}
