% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_update_map}
\alias{locationservice_update_map}
\title{Updates the specified properties of a given map resource}
\usage{
locationservice_update_map(
  MapName,
  PricingPlan = NULL,
  Description = NULL,
  ConfigurationUpdate = NULL
)
}
\arguments{
\item{MapName}{[required] The name of the map resource to update.}

\item{PricingPlan}{No longer used. If included, the only allowed value is
\code{RequestBasedUsage}.}

\item{Description}{Updates the description for the map resource.}

\item{ConfigurationUpdate}{Updates the parts of the map configuration that can be updated,
including the political view.}
}
\description{
Updates the specified properties of a given map resource.

See \url{https://www.paws-r-sdk.com/docs/locationservice_update_map/} for full documentation.
}
\keyword{internal}
