% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationservice_operations.R
\name{locationservice_update_route_calculator}
\alias{locationservice_update_route_calculator}
\title{This operation is no longer current and may be deprecated in the future}
\usage{
locationservice_update_route_calculator(
  CalculatorName,
  PricingPlan = NULL,
  Description = NULL
)
}
\arguments{
\item{CalculatorName}{[required] The name of the route calculator resource to update.}

\item{PricingPlan}{No longer used. If included, the only allowed value is
\code{RequestBasedUsage}.}

\item{Description}{Updates the description for the route calculator resource.}
}
\description{
This operation is no longer current and may be deprecated in the future. We recommend you upgrade to the Routes API V2 unless you require Grab data.

See \url{https://www.paws-r-sdk.com/docs/locationservice_update_route_calculator/} for full documentation.
}
\keyword{internal}
