% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mq_operations.R
\name{mq_delete_broker}
\alias{mq_delete_broker}
\title{Deletes a broker}
\usage{
mq_delete_broker(BrokerId)
}
\arguments{
\item{BrokerId}{[required] The name of the broker. This value must be unique in your AWS account, 1-50 characters long, must contain only letters, numbers, dashes, and underscores, and must not contain whitespaces, brackets, wildcard characters, or special characters.}
}
\description{
Deletes a broker. Note: This API is asynchronous.
}
\section{Request syntax}{
\preformatted{svc$delete_broker(
  BrokerId = "string"
)
}
}

\keyword{internal}
