% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mq_operations.R
\name{mq_describe_broker}
\alias{mq_describe_broker}
\title{Returns information about the specified broker}
\usage{
mq_describe_broker(BrokerId)
}
\arguments{
\item{BrokerId}{[required] The unique ID that Amazon MQ generates for the broker.}
}
\description{
Returns information about the specified broker.

See \url{https://www.paws-r-sdk.com/docs/mq_describe_broker/} for full documentation.
}
\keyword{internal}
