% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mq_operations.R
\name{mq_describe_broker_instance_options}
\alias{mq_describe_broker_instance_options}
\title{Describe available broker instance options}
\usage{
mq_describe_broker_instance_options(
  EngineType = NULL,
  HostInstanceType = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  StorageType = NULL
)
}
\arguments{
\item{EngineType}{Filter response by engine type.}

\item{HostInstanceType}{Filter response by host instance type.}

\item{MaxResults}{The maximum number of brokers that Amazon MQ can return per page (20 by
default). This value must be an integer from 5 to 100.}

\item{NextToken}{The token that specifies the next page of results Amazon MQ should
return. To request the first page, leave nextToken empty.}

\item{StorageType}{Filter response by storage type.}
}
\description{
Describe available broker instance options.

See \url{https://www.paws-r-sdk.com/docs/mq_describe_broker_instance_options/} for full documentation.
}
\keyword{internal}
