% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mq_operations.R
\name{mq_list_tags}
\alias{mq_list_tags}
\title{Lists tags for a resource}
\usage{
mq_list_tags(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource tag.}
}
\description{
Lists tags for a resource.

See \url{https://paws-r.github.io/docs/mq/list_tags.html} for full documentation.
}
\keyword{internal}
