% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceexplorer_operations.R
\name{resourceexplorer_associate_default_view}
\alias{resourceexplorer_associate_default_view}
\title{Sets the specified view as the default for the Amazon Web Services
Region in which you call this operation}
\usage{
resourceexplorer_associate_default_view(ViewArn)
}
\arguments{
\item{ViewArn}{[required] The \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html}{Amazon resource name (ARN)}
of the view to set as the default for the Amazon Web Services Region and
Amazon Web Services account in which you call this operation. The
specified view must already exist in the called Region.}
}
\description{
Sets the specified view as the default for the Amazon Web Services Region in which you call this operation. When a user performs a \code{\link[=resourceexplorer_search]{search}} that doesn't explicitly specify which view to use, then Amazon Web Services Resource Explorer automatically chooses this default view for searches performed in this Amazon Web Services Region.

See \url{https://www.paws-r-sdk.com/docs/resourceexplorer_associate_default_view/} for full documentation.
}
\keyword{internal}
