% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceexplorer_operations.R
\name{resourceexplorer_batch_get_view}
\alias{resourceexplorer_batch_get_view}
\title{Retrieves details about a list of views}
\usage{
resourceexplorer_batch_get_view(ViewArns = NULL)
}
\arguments{
\item{ViewArns}{A list of \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html}{Amazon resource names (ARNs)}
that identify the views you want details for.}
}
\description{
Retrieves details about a list of views.

See \url{https://www.paws-r-sdk.com/docs/resourceexplorer_batch_get_view/} for full documentation.
}
\keyword{internal}
