% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceexplorer_operations.R
\name{resourceexplorer_delete_view}
\alias{resourceexplorer_delete_view}
\title{Deletes the specified view}
\usage{
resourceexplorer_delete_view(ViewArn)
}
\arguments{
\item{ViewArn}{[required] The \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html}{Amazon resource name (ARN)}
of the view that you want to delete.}
}
\description{
Deletes the specified view.

See \url{https://www.paws-r-sdk.com/docs/resourceexplorer_delete_view/} for full documentation.
}
\keyword{internal}
