% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceexplorer_operations.R
\name{resourceexplorer_get_view}
\alias{resourceexplorer_get_view}
\title{Retrieves details of the specified view}
\usage{
resourceexplorer_get_view(ViewArn)
}
\arguments{
\item{ViewArn}{[required] The \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html}{Amazon resource name (ARN)}
of the view that you want information about.}
}
\description{
Retrieves details of the specified view.

See \url{https://www.paws-r-sdk.com/docs/resourceexplorer_get_view/} for full documentation.
}
\keyword{internal}
