% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas_operations.R
\name{schemas_describe_registry}
\alias{schemas_describe_registry}
\title{Describes the registry}
\usage{
schemas_describe_registry(RegistryName)
}
\arguments{
\item{RegistryName}{[required] The name of the registry.}
}
\description{
Describes the registry.

See \url{https://paws-r.github.io/docs/schemas/describe_registry.html} for full documentation.
}
\keyword{internal}
