% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_operations.R
\name{sfn_create_activity}
\alias{sfn_create_activity}
\title{Creates an activity}
\usage{
sfn_create_activity(name, tags = NULL, encryptionConfiguration = NULL)
}
\arguments{
\item{name}{[required] The name of the activity to create. This name must be unique for your
Amazon Web Services account and region for 90 days. For more
information, see \href{https://docs.aws.amazon.com/step-functions/latest/dg/service-quotas.html#service-limits-state-machine-executions}{Limits Related to State Machine Executions}
in the \emph{Step Functions Developer Guide}.

A name must \emph{not} contain:
\itemize{
\item white space
\item brackets \verb{< > \{ \} [ ]}
\item wildcard characters \verb{? *}
\item special characters \verb{\\" # \% \\ ^ | ~ \\` $ & , ; : /}
\item control characters (\verb{U+0000-001F}, \verb{U+007F-009F}, \code{U+FFFE-FFFF})
\item surrogates (\code{U+D800-DFFF})
\item invalid characters (\verb{ U+10FFFF})
}

To enable logging with CloudWatch Logs, the name should only contain
0-9, A-Z, a-z, - and _.

[ ]: R:\%20}

\item{tags}{The list of tags to add to a resource.

An array of key-value pairs. For more information, see \href{https://docs.aws.amazon.com/awsaccountbilling/latest/aboutv2/cost-alloc-tags.html}{Using Cost Allocation Tags}
in the \emph{Amazon Web Services Billing and Cost Management User Guide}, and
\href{https://docs.aws.amazon.com/IAM/latest/UserGuide/access_iam-tags.html}{Controlling Access Using IAM Tags}.

Tags may only contain Unicode letters, digits, white space, or these
symbols: \verb{_ . : / = + - @}.}

\item{encryptionConfiguration}{Settings to configure server-side encryption.}
}
\description{
Creates an activity. An activity is a task that you write in any programming language and host on any machine that has access to Step Functions. Activities must poll Step Functions using the \code{\link[=sfn_get_activity_task]{get_activity_task}} API action and respond using \verb{SendTask*} API actions. This function lets Step Functions know the existence of your activity and returns an identifier for use in a state machine and when polling from the activity.

See \url{https://www.paws-r-sdk.com/docs/sfn_create_activity/} for full documentation.
}
\keyword{internal}
