% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_operations.R
\name{sfn_delete_activity}
\alias{sfn_delete_activity}
\title{Deletes an activity}
\usage{
sfn_delete_activity(activityArn)
}
\arguments{
\item{activityArn}{[required] The Amazon Resource Name (ARN) of the activity to delete.}
}
\description{
Deletes an activity.
}
\section{Request syntax}{
\preformatted{svc$delete_activity(
  activityArn = "string"
)
}
}

\keyword{internal}
