% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_operations.R
\name{sfn_delete_activity}
\alias{sfn_delete_activity}
\title{Deletes an activity}
\usage{
sfn_delete_activity(activityArn)
}
\arguments{
\item{activityArn}{[required] The Amazon Resource Name (ARN) of the activity to delete.}
}
\description{
Deletes an activity.

See \url{https://paws-r.github.io/docs/sfn/delete_activity.html} for full documentation.
}
\keyword{internal}
