% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_operations.R
\name{sfn_describe_execution}
\alias{sfn_describe_execution}
\title{Describes an execution}
\usage{
sfn_describe_execution(executionArn)
}
\arguments{
\item{executionArn}{[required] The Amazon Resource Name (ARN) of the execution to describe.}
}
\description{
Describes an execution.

See \url{https://paws-r.github.io/docs/sfn/describe_execution.html} for full documentation.
}
\keyword{internal}
