% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_operations.R
\name{sfn_describe_execution}
\alias{sfn_describe_execution}
\title{Provides information about a state machine execution, such as the state
machine associated with the execution, the execution input and output,
and relevant execution metadata}
\usage{
sfn_describe_execution(executionArn, includedData = NULL)
}
\arguments{
\item{executionArn}{[required] The Amazon Resource Name (ARN) of the execution to describe.}

\item{includedData}{If your state machine definition is encrypted with a KMS key, callers
must have \code{kms:Decrypt} permission to decrypt the definition.
Alternatively, you can call DescribeStateMachine API with
\code{includedData = METADATA_ONLY} to get a successful response without the
encrypted definition.}
}
\description{
Provides information about a state machine execution, such as the state machine associated with the execution, the execution input and output, and relevant execution metadata. If you've \href{https://docs.aws.amazon.com/step-functions/latest/dg/redrive-executions.html}{redriven} an execution, you can use this API action to return information about the redrives of that execution. In addition, you can use this API action to return the Map Run Amazon Resource Name (ARN) if the execution was dispatched by a Map Run.

See \url{https://www.paws-r-sdk.com/docs/sfn_describe_execution/} for full documentation.
}
\keyword{internal}
