% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_operations.R
\name{sfn_describe_execution}
\alias{sfn_describe_execution}
\title{Provides information about a state machine execution, such as the state
machine associated with the execution, the execution input and output,
and relevant execution metadata}
\usage{
sfn_describe_execution(executionArn)
}
\arguments{
\item{executionArn}{[required] The Amazon Resource Name (ARN) of the execution to describe.}
}
\description{
Provides information about a state machine execution, such as the state machine associated with the execution, the execution input and output, and relevant execution metadata. Use this API action to return the Map Run Amazon Resource Name (ARN) if the execution was dispatched by a Map Run.

See \url{https://www.paws-r-sdk.com/docs/sfn_describe_execution/} for full documentation.
}
\keyword{internal}
