% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_operations.R
\name{sfn_describe_state_machine}
\alias{sfn_describe_state_machine}
\title{Describes a state machine}
\usage{
sfn_describe_state_machine(stateMachineArn)
}
\arguments{
\item{stateMachineArn}{[required] The Amazon Resource Name (ARN) of the state machine to describe.}
}
\description{
Describes a state machine.

See \url{https://paws-r.github.io/docs/sfn/describe_state_machine.html} for full documentation.
}
\keyword{internal}
