% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_operations.R
\name{sfn_describe_state_machine}
\alias{sfn_describe_state_machine}
\title{Provides information about a state machine's definition, its IAM role
Amazon Resource Name (ARN), and configuration}
\usage{
sfn_describe_state_machine(stateMachineArn)
}
\arguments{
\item{stateMachineArn}{[required] The Amazon Resource Name (ARN) of the state machine to describe.}
}
\description{
Provides information about a state machine's definition, its IAM role Amazon Resource Name (ARN), and configuration. If the state machine ARN is a qualified state machine ARN, the response returned includes the \code{Map} state's label.

See \url{https://www.paws-r-sdk.com/docs/sfn_describe_state_machine/} for full documentation.
}
\keyword{internal}
