% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_operations.R
\name{sfn_describe_state_machine_for_execution}
\alias{sfn_describe_state_machine_for_execution}
\title{Describes the state machine associated with a specific execution}
\usage{
sfn_describe_state_machine_for_execution(executionArn)
}
\arguments{
\item{executionArn}{[required] The Amazon Resource Name (ARN) of the execution you want state machine
information for.}
}
\description{
Describes the state machine associated with a specific execution.

See \url{https://paws-r.github.io/docs/sfn/describe_state_machine_for_execution.html} for full documentation.
}
\keyword{internal}
