% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_operations.R
\name{sfn_list_tags_for_resource}
\alias{sfn_list_tags_for_resource}
\title{List tags for a given resource}
\usage{
sfn_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) for the Step Functions state machine or
activity.}
}
\description{
List tags for a given resource.

See \url{https://paws-r.github.io/docs/sfn/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
