% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_operations.R
\name{sfn_stop_execution}
\alias{sfn_stop_execution}
\title{Stops an execution}
\usage{
sfn_stop_execution(executionArn, error, cause)
}
\arguments{
\item{executionArn}{[required] The Amazon Resource Name (ARN) of the execution to stop.}

\item{error}{The error code of the failure.}

\item{cause}{A more detailed explanation of the cause of the failure.}
}
\value{
A list with the following syntax:\preformatted{list(
  stopDate = as.POSIXct(
    "2015-01-01"
  )
)
}
}
\description{
Stops an execution.

This API action is not supported by \code{EXPRESS} state machines.
}
\section{Request syntax}{
\preformatted{svc$stop_execution(
  executionArn = "string",
  error = "string",
  cause = "string"
)
}
}

\keyword{internal}
