% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_operations.R
\name{sfn_tag_resource}
\alias{sfn_tag_resource}
\title{Add a tag to a Step Functions resource}
\usage{
sfn_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) for the Step Functions state machine or
activity.}

\item{tags}{[required] The list of tags to add to a resource.

Tags may only contain Unicode letters, digits, white space, or these
symbols: \verb{_ . : / = + - @}.}
}
\description{
Add a tag to a Step Functions resource.

See \url{https://paws-r.github.io/docs/sfn/tag_resource.html} for full documentation.
}
\keyword{internal}
