% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_operations.R
\name{sfn_validate_state_machine_definition}
\alias{sfn_validate_state_machine_definition}
\title{Validates the syntax of a state machine definition}
\usage{
sfn_validate_state_machine_definition(definition, type = NULL)
}
\arguments{
\item{definition}{[required] The Amazon States Language definition of the state machine. For more
information, see \href{https://docs.aws.amazon.com/step-functions/latest/dg/concepts-amazon-states-language.html}{Amazon States Language}
(ASL).}

\item{type}{The target type of state machine for this definition. The default is
\code{STANDARD}.}
}
\description{
Validates the syntax of a state machine definition.

See \url{https://www.paws-r-sdk.com/docs/sfn_validate_state_machine_definition/} for full documentation.
}
\keyword{internal}
