% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_operations.R
\name{sns_check_if_phone_number_is_opted_out}
\alias{sns_check_if_phone_number_is_opted_out}
\title{Accepts a phone number and indicates whether the phone holder has opted
out of receiving SMS messages from your Amazon Web Services account}
\usage{
sns_check_if_phone_number_is_opted_out(phoneNumber)
}
\arguments{
\item{phoneNumber}{[required] The phone number for which you want to check the opt out status.}
}
\description{
Accepts a phone number and indicates whether the phone holder has opted out of receiving SMS messages from your Amazon Web Services account. You cannot send SMS messages to a number that is opted out.

See \url{https://paws-r.github.io/docs/sns/check_if_phone_number_is_opted_out.html} for full documentation.
}
\keyword{internal}
