% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_operations.R
\name{sns_delete_endpoint}
\alias{sns_delete_endpoint}
\title{Deletes the endpoint for a device and mobile app from Amazon SNS}
\usage{
sns_delete_endpoint(EndpointArn)
}
\arguments{
\item{EndpointArn}{[required] EndpointArn of endpoint to delete.}
}
\description{
Deletes the endpoint for a device and mobile app from Amazon SNS. This action is idempotent. For more information, see \href{https://docs.aws.amazon.com/sns/latest/dg/sns-mobile-application-as-subscriber.html}{Using Amazon SNS Mobile Push Notifications}.

See \url{https://www.paws-r-sdk.com/docs/sns_delete_endpoint/} for full documentation.
}
\keyword{internal}
