% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_operations.R
\name{sns_get_data_protection_policy}
\alias{sns_get_data_protection_policy}
\title{Amazon SNS message data protection is no longer available to new
customers}
\usage{
sns_get_data_protection_policy(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the topic whose \code{DataProtectionPolicy} you want to get.

For more information about ARNs, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html}{Amazon Resource Names (ARNs)} in the Amazon Web Services General Reference.}
}
\description{
Amazon SNS message data protection is no longer available to new customers. For more information and guidance on alternatives, see \href{https://docs.aws.amazon.com/sns/latest/dg/sns-message-data-protection-availability-change.html}{Amazon SNS message data protection availability change}.

See \url{https://www.paws-r-sdk.com/docs/sns_get_data_protection_policy/} for full documentation.
}
\keyword{internal}
