% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_operations.R
\name{sns_get_sms_attributes}
\alias{sns_get_sms_attributes}
\title{Returns the settings for sending SMS messages from your account}
\usage{
sns_get_sms_attributes(attributes)
}
\arguments{
\item{attributes}{A list of the individual attribute names, such as \code{MonthlySpendLimit},
for which you want values.

For all attribute names, see
\href{https://docs.aws.amazon.com/sns/latest/api/API_SetSMSAttributes.html}{SetSMSAttributes}.

If you don\'t use this parameter, Amazon SNS returns all SMS attributes.}
}
\description{
Returns the settings for sending SMS messages from your account.
}
\details{
These settings are set with the \code{SetSMSAttributes} action.
}
\section{Request syntax}{
\preformatted{svc$get_sms_attributes(
  attributes = list(
    "string"
  )
)
}
}

\keyword{internal}
