% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_operations.R
\name{sns_get_sms_sandbox_account_status}
\alias{sns_get_sms_sandbox_account_status}
\title{Retrieves the SMS sandbox status for the calling Amazon Web Services
account in the target Amazon Web Services Region}
\usage{
sns_get_sms_sandbox_account_status()
}
\description{
Retrieves the SMS sandbox status for the calling Amazon Web Services account in the target Amazon Web Services Region.

See \url{https://www.paws-r-sdk.com/docs/sns_get_sms_sandbox_account_status/} for full documentation.
}
\keyword{internal}
