% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_operations.R
\name{sns_get_subscription_attributes}
\alias{sns_get_subscription_attributes}
\title{Returns all of the properties of a subscription}
\usage{
sns_get_subscription_attributes(SubscriptionArn)
}
\arguments{
\item{SubscriptionArn}{[required] The ARN of the subscription whose properties you want to get.}
}
\description{
Returns all of the properties of a subscription.

See \url{https://www.paws-r-sdk.com/docs/sns_get_subscription_attributes/} for full documentation.
}
\keyword{internal}
