% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_operations.R
\name{sns_list_platform_applications}
\alias{sns_list_platform_applications}
\title{Lists the platform application objects for the supported push
notification services, such as APNS and GCM (Firebase Cloud Messaging)}
\usage{
sns_list_platform_applications(NextToken)
}
\arguments{
\item{NextToken}{NextToken string is used when calling ListPlatformApplications action to
retrieve additional records that are available after the first page
results.}
}
\description{
Lists the platform application objects for the supported push
notification services, such as APNS and GCM (Firebase Cloud Messaging).
The results for \code{ListPlatformApplications} are paginated and return a
limited list of applications, up to 100. If additional records are
available after the first page results, then a NextToken string will be
returned. To receive the next page, you call \code{ListPlatformApplications}
using the NextToken string received from the previous call. When there
are no more records to return, \code{NextToken} will be null. For more
information, see \href{https://docs.aws.amazon.com/sns/latest/dg/SNSMobilePush.html}{Using Amazon SNS Mobile Push Notifications}.
}
\details{
This action is throttled at 15 transactions per second (TPS).
}
\section{Request syntax}{
\preformatted{svc$list_platform_applications(
  NextToken = "string"
)
}
}

\keyword{internal}
