% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_operations.R
\name{sns_list_subscriptions}
\alias{sns_list_subscriptions}
\title{Returns a list of the requester's subscriptions}
\usage{
sns_list_subscriptions(NextToken = NULL)
}
\arguments{
\item{NextToken}{Token returned by the previous
\code{\link[=sns_list_subscriptions]{list_subscriptions}} request.}
}
\description{
Returns a list of the requester's subscriptions. Each call returns a limited list of subscriptions, up to 100. If there are more subscriptions, a \code{NextToken} is also returned. Use the \code{NextToken} parameter in a new \code{\link[=sns_list_subscriptions]{list_subscriptions}} call to get further results.

See \url{https://www.paws-r-sdk.com/docs/sns_list_subscriptions/} for full documentation.
}
\keyword{internal}
