% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_operations.R
\name{sns_list_tags_for_resource}
\alias{sns_list_tags_for_resource}
\title{List all tags added to the specified Amazon SNS topic}
\usage{
sns_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the topic for which to list tags.}
}
\description{
List all tags added to the specified Amazon SNS topic. For an overview, see \href{https://docs.aws.amazon.com/sns/latest/dg/sns-tags.html}{Amazon SNS Tags} in the \emph{Amazon Simple Notification Service Developer Guide}.

See \url{https://paws-r.github.io/docs/sns/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
