% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_operations.R
\name{sns_opt_in_phone_number}
\alias{sns_opt_in_phone_number}
\title{Use this request to opt in a phone number that is opted out, which
enables you to resume sending SMS messages to the number}
\usage{
sns_opt_in_phone_number(phoneNumber)
}
\arguments{
\item{phoneNumber}{[required] The phone number to opt in.}
}
\value{
An empty list.
}
\description{
Use this request to opt in a phone number that is opted out, which
enables you to resume sending SMS messages to the number.

You can opt in a phone number only once every 30 days.
}
\section{Request syntax}{
\preformatted{svc$opt_in_phone_number(
  phoneNumber = "string"
)
}
}

\keyword{internal}
