% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_operations.R
\name{sns_publish_batch}
\alias{sns_publish_batch}
\title{Publishes up to 10 messages to the specified topic in a single batch}
\usage{
sns_publish_batch(TopicArn, PublishBatchRequestEntries)
}
\arguments{
\item{TopicArn}{[required] The Amazon resource name (ARN) of the topic you want to batch publish
to.}

\item{PublishBatchRequestEntries}{[required] A list of \code{\link[=sns_publish_batch]{publish_batch}} request entries to be
sent to the SNS topic.}
}
\description{
Publishes up to 10 messages to the specified topic in a single batch. This is a batch version of the \code{\link[=sns_publish]{publish}} API. If you try to send more than 10 messages in a single batch request, you will receive a \code{TooManyEntriesInBatchRequest} exception.

See \url{https://www.paws-r-sdk.com/docs/sns_publish_batch/} for full documentation.
}
\keyword{internal}
