% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_operations.R
\name{sns_put_data_protection_policy}
\alias{sns_put_data_protection_policy}
\title{Amazon SNS message data protection is no longer available to new
customers}
\usage{
sns_put_data_protection_policy(ResourceArn, DataProtectionPolicy)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the topic whose \code{DataProtectionPolicy} you want to add or update.

For more information about ARNs, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html}{Amazon Resource Names (ARNs)} in the Amazon Web Services General Reference.}

\item{DataProtectionPolicy}{[required] The JSON serialization of the topic's \code{DataProtectionPolicy}.

The \code{DataProtectionPolicy} must be in JSON string format.

Length Constraints: Maximum length of 30,720.}
}
\description{
Amazon SNS message data protection is no longer available to new customers. For more information and guidance on alternatives, see \href{https://docs.aws.amazon.com/sns/latest/dg/sns-message-data-protection-availability-change.html}{Amazon SNS message data protection availability change}.

See \url{https://www.paws-r-sdk.com/docs/sns_put_data_protection_policy/} for full documentation.
}
\keyword{internal}
