% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sns_operations.R
\name{sns_put_data_protection_policy}
\alias{sns_put_data_protection_policy}
\title{Adds or updates an inline policy document that is stored in the
specified Amazon SNS topic}
\usage{
sns_put_data_protection_policy(ResourceArn, DataProtectionPolicy)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the topic whose \code{DataProtectionPolicy} you want to add or
update.

For more information about ARNs, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html}{Amazon Resource Names (ARNs)}
in the Amazon Web Services General Reference.}

\item{DataProtectionPolicy}{[required] The JSON serialization of the topic's \code{DataProtectionPolicy}.

The \code{DataProtectionPolicy} must be in JSON string format.

Length Constraints: Maximum length of 30,720.}
}
\description{
Adds or updates an inline policy document that is stored in the specified Amazon SNS topic.

See \url{https://www.paws-r-sdk.com/docs/sns_put_data_protection_policy/} for full documentation.
}
\keyword{internal}
