% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqs_operations.R
\name{sqs_get_queue_url}
\alias{sqs_get_queue_url}
\title{Returns the URL of an existing Amazon SQS queue}
\usage{
sqs_get_queue_url(QueueName, QueueOwnerAWSAccountId = NULL)
}
\arguments{
\item{QueueName}{[required] The name of the queue whose URL must be fetched. Maximum 80 characters.
Valid values: alphanumeric characters, hyphens (\code{-}), and underscores
(\verb{_}).

Queue URLs and names are case-sensitive.}

\item{QueueOwnerAWSAccountId}{The Amazon Web Services account ID of the account that created the
queue.}
}
\description{
Returns the URL of an existing Amazon SQS queue.

See \url{https://paws-r.github.io/docs/sqs/get_queue_url.html} for full documentation.
}
\keyword{internal}
