% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqs_operations.R
\name{sqs_list_dead_letter_source_queues}
\alias{sqs_list_dead_letter_source_queues}
\title{Returns a list of your queues that have the RedrivePolicy queue
attribute configured with a dead-letter queue}
\usage{
sqs_list_dead_letter_source_queues(
  QueueUrl,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{QueueUrl}{[required] The URL of a dead-letter queue.

Queue URLs and names are case-sensitive.}

\item{NextToken}{Pagination token to request the next set of results.}

\item{MaxResults}{Maximum number of results to include in the response. Value range is 1
to 1000. You must set \code{MaxResults} to receive a value for \code{NextToken} in
the response.}
}
\description{
Returns a list of your queues that have the \code{RedrivePolicy} queue attribute configured with a dead-letter queue.

See \url{https://www.paws-r-sdk.com/docs/sqs_list_dead_letter_source_queues/} for full documentation.
}
\keyword{internal}
