% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqs_operations.R
\name{sqs_list_queue_tags}
\alias{sqs_list_queue_tags}
\title{List all cost allocation tags added to the specified Amazon SQS queue}
\usage{
sqs_list_queue_tags(QueueUrl)
}
\arguments{
\item{QueueUrl}{[required] The URL of the queue.}
}
\description{
List all cost allocation tags added to the specified Amazon SQS queue. For an overview, see \href{https://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/sqs-queue-tags.html}{Tagging Your Amazon SQS Queues} in the \emph{Amazon SQS Developer Guide}.

See \url{https://paws-r.github.io/docs/sqs/list_queue_tags.html} for full documentation.
}
\keyword{internal}
