% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqs_operations.R
\name{sqs_remove_permission}
\alias{sqs_remove_permission}
\title{Revokes any permissions in the queue policy that matches the specified
Label parameter}
\usage{
sqs_remove_permission(QueueUrl, Label)
}
\arguments{
\item{QueueUrl}{[required] The URL of the Amazon SQS queue from which permissions are removed.

Queue URLs and names are case-sensitive.}

\item{Label}{[required] The identification of the permission to remove. This is the label added
using the \code{\link[=sqs_add_permission]{add_permission}} action.}
}
\description{
Revokes any permissions in the queue policy that matches the specified \code{Label} parameter.

See \url{https://paws-r.github.io/docs/sqs/remove_permission.html} for full documentation.
}
\keyword{internal}
