% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqs_operations.R
\name{sqs_set_queue_attributes}
\alias{sqs_set_queue_attributes}
\title{Sets the value of one or more queue attributes, like a policy}
\usage{
sqs_set_queue_attributes(QueueUrl, Attributes)
}
\arguments{
\item{QueueUrl}{[required] The URL of the Amazon SQS queue whose attributes are set.

Queue URLs and names are case-sensitive.}

\item{Attributes}{[required] A map of attributes to set.

The following lists the names, descriptions, and values of the special
request parameters that the
\code{\link[=sqs_set_queue_attributes]{set_queue_attributes}} action uses:
\itemize{
\item \code{DelaySeconds} – The length of time, in seconds, for which the
delivery of all messages in the queue is delayed. Valid values: An
integer from 0 to 900 (15 minutes). Default: 0.
\item \code{MaximumMessageSize} – The limit of how many bytes a message can
contain before Amazon SQS rejects it. Valid values: An integer from
1,024 bytes (1 KiB) up to 1,048,576 bytes (1 MiB). Default:
1,048,576 bytes (1 MiB).
\item \code{MessageRetentionPeriod} – The length of time, in seconds, for which
Amazon SQS retains a message. Valid values: An integer representing
seconds, from 60 (1 minute) to 1,209,600 (14 days). Default: 345,600
(4 days). When you change a queue's attributes, the change can take
up to 60 seconds for most of the attributes to propagate throughout
the Amazon SQS system. Changes made to the \code{MessageRetentionPeriod}
attribute can take up to 15 minutes and will impact existing
messages in the queue potentially causing them to be expired and
deleted if the \code{MessageRetentionPeriod} is reduced below the age of
existing messages.
\item \code{Policy} – The queue's policy. A valid Amazon Web Services policy.
For more information about policy structure, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/access_policies.html}{Overview of Amazon Web Services IAM Policies}
in the \emph{Identity and Access Management User Guide}.
\item \code{ReceiveMessageWaitTimeSeconds} – The length of time, in seconds,
for which a \code{\link[=sqs_receive_message]{receive_message}} action waits
for a message to arrive. Valid values: An integer from 0 to 20
(seconds). Default: 0.
\item \code{VisibilityTimeout} – The visibility timeout for the queue, in
seconds. Valid values: An integer from 0 to 43,200 (12 hours).
Default: 30. For more information about the visibility timeout, see
\href{https://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/sqs-visibility-timeout.html}{Visibility Timeout}
in the \emph{Amazon SQS Developer Guide}.
}

The following attributes apply only to \href{https://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/sqs-dead-letter-queues.html}{dead-letter queues:}
\itemize{
\item \code{RedrivePolicy} – The string that includes the parameters for the
dead-letter queue functionality of the source queue as a JSON
object. The parameters are as follows:
\itemize{
\item \code{deadLetterTargetArn} – The Amazon Resource Name (ARN) of the
dead-letter queue to which Amazon SQS moves messages after the
value of \code{maxReceiveCount} is exceeded.
\item \code{maxReceiveCount} – The number of times a message is delivered
to the source queue before being moved to the dead-letter queue.
Default: 10. When the \code{ReceiveCount} for a message exceeds the
\code{maxReceiveCount} for a queue, Amazon SQS moves the message to
the dead-letter-queue.
}
\item \code{RedriveAllowPolicy} – The string that includes the parameters for
the permissions for the dead-letter queue redrive permission and
which source queues can specify dead-letter queues as a JSON object.
The parameters are as follows:
\itemize{
\item \code{redrivePermission} – The permission type that defines which
source queues can specify the current queue as the dead-letter
queue. Valid values are:
\itemize{
\item \code{allowAll} – (Default) Any source queues in this Amazon Web
Services account in the same Region can specify this queue
as the dead-letter queue.
\item \code{denyAll} – No source queues can specify this queue as the
dead-letter queue.
\item \code{byQueue} – Only queues specified by the \code{sourceQueueArns}
parameter can specify this queue as the dead-letter queue.
}
\item \code{sourceQueueArns} – The Amazon Resource Names (ARN)s of the
source queues that can specify this queue as the dead-letter
queue and redrive messages. You can specify this parameter only
when the \code{redrivePermission} parameter is set to \code{byQueue}. You
can specify up to 10 source queue ARNs. To allow more than 10
source queues to specify dead-letter queues, set the
\code{redrivePermission} parameter to \code{allowAll}.
}
}

The dead-letter queue of a FIFO queue must also be a FIFO queue.
Similarly, the dead-letter queue of a standard queue must also be a
standard queue.

The following attributes apply only to
\href{https://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/sqs-server-side-encryption.html}{server-side-encryption}:
\itemize{
\item \code{KmsMasterKeyId} – The ID of an Amazon Web Services managed customer
master key (CMK) for Amazon SQS or a custom CMK. For more
information, see \href{https://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/sqs-server-side-encryption.html#sqs-sse-key-terms}{Key Terms}.
While the alias of the AWS-managed CMK for Amazon SQS is always
\code{alias/aws/sqs}, the alias of a custom CMK can, for example, be
\code{alias/MyAlias }. For more examples, see
\href{https://docs.aws.amazon.com/kms/latest/APIReference/API_DescribeKey.html#API_DescribeKey_RequestParameters}{KeyId}
in the \emph{Key Management Service API Reference}.
\item \code{KmsDataKeyReusePeriodSeconds} – The length of time, in seconds, for
which Amazon SQS can reuse a \href{https://docs.aws.amazon.com/kms/latest/developerguide/concepts.html#data-keys}{data key}
to encrypt or decrypt messages before calling KMS again. An integer
representing seconds, between 60 seconds (1 minute) and 86,400
seconds (24 hours). Default: 300 (5 minutes). A shorter time period
provides better security but results in more calls to KMS which
might incur charges after Free Tier. For more information, see \href{https://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/sqs-server-side-encryption.html#sqs-how-does-the-data-key-reuse-period-work}{How Does the Data Key Reuse Period Work?}.
\item \code{SqsManagedSseEnabled} – Enables server-side queue encryption using
SQS owned encryption keys. Only one server-side encryption option is
supported per queue (for example,
\href{https://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/sqs-configure-sse-existing-queue.html}{SSE-KMS}
or
\href{https://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/sqs-configure-sqs-sse-queue.html}{SSE-SQS}).
}

The following attribute applies only to \href{https://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/}{FIFO (first-in-first-out) queues}:
\itemize{
\item \code{ContentBasedDeduplication} – Enables content-based deduplication.
For more information, see \href{https://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/FIFO-queues-exactly-once-processing.html}{Exactly-once processing}
in the \emph{Amazon SQS Developer Guide}. Note the following:
\itemize{
\item Every message must have a unique \code{MessageDeduplicationId}.
\itemize{
\item You may provide a \code{MessageDeduplicationId} explicitly.
\item If you aren't able to provide a \code{MessageDeduplicationId} and
you enable \code{ContentBasedDeduplication} for your queue,
Amazon SQS uses a SHA-256 hash to generate the
\code{MessageDeduplicationId} using the body of the message (but
not the attributes of the message).
\item If you don't provide a \code{MessageDeduplicationId} and the
queue doesn't have \code{ContentBasedDeduplication} set, the
action fails with an error.
\item If the queue has \code{ContentBasedDeduplication} set, your
\code{MessageDeduplicationId} overrides the generated one.
}
\item When \code{ContentBasedDeduplication} is in effect, messages with
identical content sent within the deduplication interval are
treated as duplicates and only one copy of the message is
delivered.
\item If you send one message with \code{ContentBasedDeduplication} enabled
and then another message with a \code{MessageDeduplicationId} that is
the same as the one generated for the first
\code{MessageDeduplicationId}, the two messages are treated as
duplicates and only one copy of the message is delivered.
}
}

The following attributes apply only to \href{https://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/high-throughput-fifo.html}{high throughput for FIFO queues}:
\itemize{
\item \code{DeduplicationScope} – Specifies whether message deduplication
occurs at the message group or queue level. Valid values are
\code{messageGroup} and \code{queue}.
\item \code{FifoThroughputLimit} – Specifies whether the FIFO queue throughput
quota applies to the entire queue or per message group. Valid values
are \code{perQueue} and \code{perMessageGroupId}. The \code{perMessageGroupId}
value is allowed only when the value for \code{DeduplicationScope} is
\code{messageGroup}.
}

To enable high throughput for FIFO queues, do the following:
\itemize{
\item Set \code{DeduplicationScope} to \code{messageGroup}.
\item Set \code{FifoThroughputLimit} to \code{perMessageGroupId}.
}

If you set these attributes to anything other than the values shown for
enabling high throughput, normal throughput is in effect and
deduplication occurs as specified.

For information on throughput quotas, see \href{https://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/quotas-messages.html}{Quotas related to messages}
in the \emph{Amazon SQS Developer Guide}.}
}
\description{
Sets the value of one or more queue attributes, like a policy. When you change a queue's attributes, the change can take up to 60 seconds for most of the attributes to propagate throughout the Amazon SQS system. Changes made to the \code{MessageRetentionPeriod} attribute can take up to 15 minutes and will impact existing messages in the queue potentially causing them to be expired and deleted if the \code{MessageRetentionPeriod} is reduced below the age of existing messages.

See \url{https://www.paws-r-sdk.com/docs/sqs_set_queue_attributes/} for full documentation.
}
\keyword{internal}
