% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqs_operations.R
\name{sqs_tag_queue}
\alias{sqs_tag_queue}
\title{Add cost allocation tags to the specified Amazon SQS queue}
\usage{
sqs_tag_queue(QueueUrl, Tags)
}
\arguments{
\item{QueueUrl}{[required] The URL of the queue.}

\item{Tags}{[required] The list of tags to be added to the specified queue.}
}
\description{
Add cost allocation tags to the specified Amazon SQS queue. For an overview, see \href{https://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/sqs-queue-tags.html}{Tagging Your Amazon SQS Queues} in the \emph{Amazon SQS Developer Guide}.

See \url{https://paws-r.github.io/docs/sqs/tag_queue.html} for full documentation.
}
\keyword{internal}
