% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf_operations.R
\name{swf_deprecate_workflow_type}
\alias{swf_deprecate_workflow_type}
\title{Deprecates the specified workflow type}
\usage{
swf_deprecate_workflow_type(domain, workflowType)
}
\arguments{
\item{domain}{[required] The name of the domain in which the workflow type is registered.}

\item{workflowType}{[required] The workflow type to deprecate.}
}
\description{
Deprecates the specified \emph{workflow type}. After a workflow type has been deprecated, you cannot create new executions of that type. Executions that were started before the type was deprecated continues to run. A deprecated workflow type may still be used when calling visibility actions.

See \url{https://paws-r.github.io/docs/swf/deprecate_workflow_type.html} for full documentation.
}
\keyword{internal}
