% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf_operations.R
\name{swf_describe_workflow_type}
\alias{swf_describe_workflow_type}
\title{Returns information about the specified workflow type}
\usage{
swf_describe_workflow_type(domain, workflowType)
}
\arguments{
\item{domain}{[required] The name of the domain in which this workflow type is registered.}

\item{workflowType}{[required] The workflow type to describe.}
}
\description{
Returns information about the specified \emph{workflow type}. This includes configuration settings specified when the type was registered and other information such as creation date, current status, etc.

See \url{https://paws-r.github.io/docs/swf/describe_workflow_type.html} for full documentation.
}
\keyword{internal}
