% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf_operations.R
\name{swf_list_tags_for_resource}
\alias{swf_list_tags_for_resource}
\title{List tags for a given domain}
\usage{
swf_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) for the Amazon SWF domain.}
}
\description{
List tags for a given domain.

See \url{https://www.paws-r-sdk.com/docs/swf_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
