% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf_operations.R
\name{swf_tag_resource}
\alias{swf_tag_resource}
\title{Add a tag to a Amazon SWF domain}
\usage{
swf_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) for the Amazon SWF domain.}

\item{tags}{[required] The list of tags to add to a domain.

Tags may only contain unicode letters, digits, whitespace, or these
symbols: \verb{_ . : / = + - @}.}
}
\description{
Add a tag to a Amazon SWF domain.

See \url{https://paws-r.github.io/docs/swf/tag_resource.html} for full documentation.
}
\keyword{internal}
