% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf_operations.R
\name{swf_untag_resource}
\alias{swf_untag_resource}
\title{Remove a tag from a Amazon SWF domain}
\usage{
swf_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) for the Amazon SWF domain.}

\item{tagKeys}{[required] The list of tags to remove from the Amazon SWF domain.}
}
\description{
Remove a tag from a Amazon SWF domain.

See \url{https://paws-r.github.io/docs/swf/untag_resource.html} for full documentation.
}
\keyword{internal}
